==> code/REVISIONS/7_GridData_2022.Rmd <==
---
title: "Apply Grid to Data"
author: "Yvonne Barkley"
date: "6/30/2020"
output: html_notebook
editor_options: 
  chunk_output_type: console
---
This code creates the grid and centroids used for calculating the sampling units for the models to incorporate passive acoustic data for any species.

Technical Note: All grids and centroids are calculated separately for each survey to more easily accommodate the different timing and trackline data. However, a more recent change is to combine all of teh centroid data for processing the environmental data for each model type (Combined, Acoustics Only, etc). This may take longer, but it's much easier than splitting up the centroids by survey on separate computers and then having to keep track of everything and consolidate it later, ugh. IT'S MUCH EASIER NOW. 
That said, a loop to load all centroid data is located in the LAST CHUNK of the markdown.


Load functions 
```{r}

source(here::here('./code/FromTaiki/gridFunctions_v3.R'))
library(PAMmisc)
library(tidyverse)
library(lubridate)
library(Distance)
```

##Load parameters
```{r}
surveynum = 1706
gridsize = 16
loctype = 'Combined'
loctype2 = 'Comb'

```

##Load data if necessary
To revisit a data set:
*load grid or centroid data
*run GPS data again
*filter Pm detections for survey
*get centroids if needed
Then proceed to 'GRID QC' chunk below.
```{r}
#load grid data
gridEffortOLD <- readRDS(here::here( paste0('output/gridEffort/', gridsize, ' km-', loctype, '/gridEffort', surveynum, '_', gridsize, 'km_', loctype2, '.rda') ))

gridEffortNEW <- readRDS(here::here(paste0('output/gridEffort/', gridsize, ' km-', loctype), paste0('gridEffort', surveynum, '_', gridsize, 'km_', loctype2, 'NEW.rda')))
gridEffortOLD <- readRDS(here::here(paste0('output/gridEffort/', gridsize, ' km-', loctype), paste0('gridEffort', surveynum, '_', gridsize, 'km_', loctype2, 'OLD.rda')))
gridEffort <- readRDS(here::here( paste0('output/gridEffort/', gridsize, ' km-', loctype), paste0('gridEffort', surveynum, '_', gridsize, 'km_', loctype2, '_raw.rda') ))


#load centroid locations
centroidAll <- readRDS(here::here(paste0('output/centroids/', gridsize, ' km-', loctype), paste0('centroids',  surveynum, '_', gridsize, 'km_', loctype2, '_raw.rda')))

```

##1. GPS data
Edited 1641 to include headDiff1 as 0 for the timeframe associated with A326 I think. Not sure if that was necessary.  All other surveys should be left unedited.
```{r}
# EXAMPLE WORK FLOW
# 1) First read in gps, then you need to convert longitudes to 0-360,
# convert UTC to POSIXct format, and create a column called "effort" that is TRUE or FALSE
# Column names needed: Longitude, Latitude, UTC, effort
gpsAll <- read.csv(paste0('./output/straightPath results/straightPathWeffort_', surveynum, '.csv'), stringsAsFactors = FALSE)
### gpsAll <- read.csv(paste0('./output/straightPath results/straightPathWeffort_', surveynum, '_edit.csv'), stringsAsFactors = FALSE)  #edited 1641 but the UTC dropped the seconds ugh, so replaced it here:

gpsAll$UTC <- lubridate::ymd_hms(gpsAll$UTC)
# gpsAll$UTC <- gpsAll2$UTC  #used for 1641 only

unique(gpsAll$join.veffort)
unique(gpsAll$join.aeffort)
#TRIM WHITE SPACE IF NEEDED
gpsAll$join.veffort <- str_trim(gpsAll$join.veffort, side = "right")
gpsAll$join.aeffort <- str_trim(gpsAll$join.aeffort, side = "right")


# gpsAll <- gpsAll %>% mutate_if(is.character, str_trim) #white space after join.veffort!?!?
# gpsAll$effort <- gpsAll$straight == TRUE & (gpsAll$join.aeffort == "on" | gpsAll$join.veffort =="on") # effort used for 'allsw'
gpsAll$effort <- gpsAll$straight == TRUE & gpsAll$join.aeffort == "on" & gpsAll$join.veffort =="on"  # COMBINED EFFORT: Visual and Acoustics BOTH must be on

gpsAll$Longitude <- ifelse(gpsAll$Longitude <= 0, gpsAll$Longitude + 360, gpsAll$Longitude)

```

```{r}
#EFFORT CHECKING

# Both should be zero for the combined model
#when vis is ON and acoustics is OFF, effort should be FALSE
effVischk=gpsAll[which(gpsAll$join.veffort=='on' & gpsAll$join.aeffort =='off' & gpsAll$straight==TRUE), ]
which(effVischk$effort == TRUE) #should be zero, don't want TRUE effort if ac is off

#when vis is OFF and acoustics is OFF, effort should still be FALSE
effACchk=gpsAll[which(gpsAll$join.veffort=='off' &  gpsAll$join.aeffort=='on'   &  gpsAll$straight==TRUE),]
which(effACchk$effort == TRUE) #should be zero, Don't want this to be false
#Both outputs should be zero

effTotChk=gpsAll[which(gpsAll$straight==TRUE, (gpsAll$join.veffort=='on' &  gpsAll$join.aeffort=='on')), ]


```

###subset GPS for troubleshooting
```{r}

gpsAll <- filter(gpsAll, Id >= 146246 & Id <= 148253)

```



##2. Load sperm whale detection data
Includes localized acoustic encounters and sightings with acoustics only.
```{r}
# 2) Read in detections, convert column names to Longitude, Latitude, UTC,
# and "distance" if you need to create a detection function using Distance package
# Longitude must be matching 0-360 like in the GPS data

pmDets <- read.csv('./output/SpermiesFinalLR_20200919.csv', stringsAsFactors = FALSE)
# pmDets <- filter(SwEnvFinal, loc == 1 | loc == 0 & sid < 999) # localized encs and sightings

#### This is where you change the location from localized position to trackline position ####
pmDets <- rename(pmDets, Longitude =lon2, #lon_trk2
                         Latitude=lat, #lat_trk
                         distance = pdist)

pmDets$Longitude <- ifelse(pmDets$Longitude <= 0, pmDets$Longitude + 360, pmDets$Longitude)
pmDets$UTC <- lubridate::mdy_hm(pmDets$UTC)
pmDets$distance <- abs(pmDets$distance)*1000


# pmDets <- filter(pmDets, type == 'best' | sid < 999, itrk==1)
pmDetsAc <- filter(pmDets, sid == 999)
# pmDetsub <- filter(pmDets, survey == surveynum, acid >= 71 & acid <= 74)
```

##Make detection function
```{r}
# 3) fit a detection function

#winner! See v2 for other variations
dsmAll_3pcthn <- Distance::ds(pmDetsAc, key = 'hn', truncation = '3%')

plot(dsmAll_3pcthn)

summary(dsmAll_3pcthn)
gof_ds(dsmAll_3pcthn)

detfxn <- dsmAll_3pcthn
summ <- summary(detfxn)

```

##Calculate BIG Grid 
```{r}

# 4) Run grid stuff. This will take a while, and should have a couple progress bars showing you how long the slow parts
# are taking. Set "trunc_m" to whatever truncation distance you might want to use
# It will ask you what pixel size you want  (this is the grid size), in the future you can specify this when you call it

# YB Sept 3, 2020: Make overall larger grid that can be reused for each spatial scale

bigGrid <- doAllGrid(gps = gpsAll,
                     bounds = NULL,
                     dets = pmDets[, c(9,25)],
                     trunc_m = 20e3, #METERS
                     dsmodel = NULL, #NULL or dsm 
                     pixel = 0.144,  # 16 km
                     grid = NULL,
                     plot = F)
#Save raw OVERALL grid 
saveRDS(bigGrid, paste0('output/gridEffort/bigGrids/', 'GRID', surveynum, '_', gridsize,'km_', loctype2, '.rda') )

# LOAD OVERALL GRID IF NEEDED 
# This is the overall grid - use 25km trunc_m
# bigGrid <- readRDS( paste0('output/gridEffort/bigGrids/', 'GRID', surveynum, '_', gridsize,'km_', loctype2, '.rda') )
bigGrid <- readRDS( paste0('output/gridEffort/bigGrids/', 'GRID', surveynum, '_', gridsize,'km.rda') )

```


##Run Effort and Centroids

Centroids are only calculated for cells with effort. Sometimes, a cell with a detection ends up with no effort. It's possible to manually enter it into `gridEffort$effort$area` to override the function using the gridIndex and basing the effort on the actualArea value in the detections df. 

```{r}

# FILTER whales by survey 
# pmDetsub <- dplyr::filter(pmDets, survey == surveynum, loc == 1 | loc == 0 & sid < 999 ) # localized encs and sightings
# pmDetsub <- dplyr::filter(pmDets, survey == surveynum) #include only whales for the survey here, then they will be combined later 

#### !!!! 2022 Change to Detections !!!! ####
## To allow for random selection of localized encounters, include ALL DETECTIONS in the gridding process that can be filtered later
## Click codes were changed in SpermiesWithEnvData_20220324
#March 2022
# pmDets <- read.csv('./output/SpermiesWithEnvData_20220324_editdCodes.csv', stringsAsFactors = FALSE)

#April 2022
pmDets <- read.csv('./output/SpermiesWithEnvData_20220418_edited.csv', stringsAsFactors = FALSE)

pmDets <- rename(pmDets, Longitude =lon2, #lon_trk2
                         Latitude=lat, #lat_trk
                         distance = pdist)

# pmDets$Longitude <- ifelse(pmDets$Longitude <= 0, pmDets$Longitude + 360, pmDets$Longitude)
pmDets$UTC <- lubridate::mdy_hm(pmDets$UTC)
pmDets$distance <- abs(pmDets$distance)*1000


# 2022: Includes both peaks A and B for 'best' localized positions, sightings, and non-localized acoustic encounters
pmDetsub <- filter(pmDets, survey == surveynum, type=='best' & itrk ==1 | loc == 0 & acid < 999)  

gridEffort <- doAllGrid(gps = gpsAll,
                     bounds = NULL,
                     dets = pmDetsub,
                     trunc_m = 16215, #as.numeric(summ$ds$width[1]), #METERS
                     dsmodel = detfxn, #NULL or dsm 
                     pixel = 0.144,
                     grid = bigGrid$grid,  ###!! Use the correct big grid !!!
                     plot = F)

# Result of this is a list with the gps you provided, the grid it created, the effort values for each grid, and the detections you provided (these have the effortArea attached to them, as well as the actualArea of that grid cell - actualArea is mostly there as a reference point for sanity check, prob not used for much)

#The centroid function has to be run twice if there are detections with no effort.
#function calculates the difference between start and end times in minutes
centroidAll <- gridCentroid(gridEffort, gpsAll) #assigns the 'pa' column where 0 is no whale, 1 is whale

# centroidAll2<- gridCentroid(gridEffort2, gpsAll)


#Save raw gridEffort and centroid data

saveRDS(gridEffort, here::here(  paste0('output/gridEffort/', gridsize, ' km-', loctype, '/gridEffort', surveynum, '_', gridsize, 'km_', loctype2, '_raw.rda') ))

saveRDS(centroidAll, here::here(paste0('output/centroids/', gridsize, ' km-', loctype), paste0('centroids', surveynum, '_', gridsize, 'km_', loctype2, '_raw.rda') ))
```


#GRID QC
All detections should be included and each have effort > 0. Cells with zero effort cannot have centroids calculated. If all is well here, then proceed to centroids.
Otherwise, make changes to any detections in 'Resolve cell...' chunk.
```{r}
#1. check that all detections are included as expected. 
(pmDetsub$ID %in% gridEffort$detections$ID) #Should be all TRUE
data.frame(pmDetsub$ID , gridEffort$detections$ID) #if mismatch, this won't work
# (gridEffort$detections$ID %in% pmDetsub$ID) #Should be all TRUE

#2. which detections have effortArea = 0? (Which are deemed to have 0 effort? No bueno.)
(effdets <- which(gridEffort$detections$effortArea==0)) #output is the index
(effchg <- gridEffort$detections[effdets,]) #output is all info for each detection with 0 effort

#3. check for duplicate detections in the same cells
(gridIndex <- as.data.frame(table(gridEffort$detections$gridIndex)) %>%
    rename(gridIndex = Var1, frequency = Freq) %>%
  arrange(desc(frequency)))


if (length(effdets) == 0 & nrow(effchg) == 0){
print('All detections are included and have effort')
} else {
  print('Some detections missing effort')
}

```


#IF CELLS WITH WHALES EXIST WITH NO EFFORT
!!  After making any changes to effort, it's good to run the centroids again  !!
Do the following steps.
##1. Resolve whale cell(s) with no effort 
```{r}
#Map out the surrounding cells to find ones with effort nearby
gridmap <- attr(gridEffort$grid, 'map')

#Sometimes >1 detections have 0 effort, can look at all of them
# g = 1 # which row in effchg?
# GridInd = effchg[g,]$gridIndex # what grid cell is it in?
GridInd = effchg[,c(3,43)] # pull acid and gridIndex for all detections with no effort
gridEffort$effort$area[GridInd[,2]] # how much effort is in the cell? (Should be zero)

#Get cell index for the closest surrounding cells. Use stack to make one column for them.t means transpose
GridSurroundtmp <- data.frame( t( gridmap[GridInd[,2],] ) )
colnames(GridSurroundtmp) <- GridInd[,1]
GridSurround <- stack(GridSurroundtmp) #output 2 columns, 1: grid cell, 2: ac. ID
# GridSurround$acid <- GridInd[,1]       #add column for ac ID
# GridSurround <- GridSurround[,-2]      #remove 'ind' col of X1...X4 col names

GridEff <- data.frame(gridEffort$effort$area[GridSurround$values]) # get effort in surr cells
GridDir <- c('S', 'E', 'N', 'W') #make this vector
GridSurrDF <- data.frame(GridSurround, GridEff, GridDir) #combine surr cells and their effort and direction
colnames(GridSurrDF) <- c("gridID", "acid", "effort", "Dir")
GridSurrDF

#Get the centroid locations of the surrounding cells that have effort around the cell for the detection that has no effort.
CentSurround <- filter(centroidAll, effCells %in% GridSurrDF$gridID) #these cells all have effort

#JOIN centroids with grid map cells info and drop NAs (cells without effort)
GridCentJoin <- GridSurrDF %>% full_join(CentSurround, by = c("gridID" = "effCells")) 
GridCentJoin<-GridCentJoin %>% drop_na()
```



##2. Plot grid of whale cell(s) with no effort
REMEMBER: GridCentJoin only includes the surrounding cells, 'Dir', WITH effort, so not all 'Dir' directions may be included in the plot.
CHOOSE the replacement cells using the direction, GridCentJoin$Dir.
```{r}
x <- gridEffort
actualArea <- as.numeric(st_area(x$grid))
coveragePct <- round(x$effort$area / actualArea, 3)
coveragePct <- ifelse(coveragePct > 1, 1, coveragePct)

pdf(here::here('figures', paste0('Grid', surveynum, '_', gridsize, 'km_', loctype, '_noEffort_chk2.pdf')), width=11,height=8)

plot(x$grid, col = gray(1 - coveragePct, alpha = .9), lwd = 0.1)
points(x=x$gps$Longitude, y=x$gps$Latitude, col='blue', pch=15, cex = 0.05)
# points(x=pmDetsub$Longitude, y=pmDetsub$Latitude, col='green', pch=1, cex=0.1, lwd=0.1)
# text(pmDetsub$Longitude, pmDetsub$Latitude,col='magenta', labels=pmDetsub$acid, cex=0.1, lwd = 0.1 )

# points(x=gridEffort$detections$Longitude, y=gridEffort$detections$Latitude, col='green', pch=10, cex=0.1, lwd=0.1)
# text(gridEffort$detections$Longitude, gridEffort$detections$Latitude,col='magenta', labels=gridEffort$detections$acid, cex=0.2, lwd = 0.1 )
points(centroidAll$lon, centroidAll$lat, col='red', pch = 19, cex=0.05, lwd=0.1)

# points(x=pmDetsub$Longitude, y=pmDetsub$Latitude, col='green', pch=0, cex=0.2, lwd=0.1)
# points(	 178.0128, 22.57317, col='darkorchid', pch=8, cex=0.1, lwd=0.1) #grid locations of 
# # text(	191.1134, 24.82075,col='black', labels=pmDetsGrid1705$acid, cex=0.1, lwd = 0.1 )
# points(	192.0295, 21.76568, col='darkorange', pch=8, cex=0.1, lwd=0.1) #grid locations of 
# text(	192.0295, 	21.76568, col='black', labels=pmDetsGrid1705$acid, cex=0.1, lwd = 0.1 )
# points(192.7076, 24.31634, col='magenta', pch = 0, cex=0.2, lwd=0.1) #where is this centroid compared to the actual location? for 1303 sighting to change depth to -697
# points(192.7200, 24.28700, col='goldenrod', pch = 0, cex=0.2, lwd=0.1) #where is this centroid compared to the actual location?

#plot detections and their centroids missing effort
text(x=effchg$Longitude, y=effchg$Latitude	, col='magenta', labels=effchg$acid, cex=0.3, lwd = 0.1 )
points(x=effchg$Longitude, y=effchg$Latitude	, col='orange', pch = 19, cex=0.1, lwd = 0.07 )
text(GridCentJoin$lon, GridCentJoin$lat, col='darkviolet', labels=GridCentJoin$Dir, cex=0.1, lwd=0.1)

#1705 for A182 which had the wrong lat/lon
# points(187.69, 22.666, col= 'black', pch = 8, cex=0.2, lwd = 0.07)


#plot detections missing in the centroid data
# text(x=CentMiss$Longitude, y=CentMiss$Latitude	, col='magenta', labels=CentMiss$acid, cex=0.1, lwd = 0.1 )
#  points(x=CentMiss$Longitude, y=CentMiss$Latitude	, col='orange', pch = 19, cex=0.07, lwd = 0.07 )

dev.off()


```


##3. Replace whale cells with no effort
2022 - OR do the steps below
Another way to resolve effort, use the actualArea to get the equivalent value of effortArea
```{r}
#1641: the A326S242 SHOULD have effort calculated since both were ON along straight trackline = 134387571, gridIndex=17636
#1604: A28S999 SHOULD have effort = 20425116, gridIndex=10406


eff = 3217647    # this is the effort value to substitute into the grid with zero effort
ind = 2
grid_chg = gridEffort$detections[effdets[ind], 43]  #this is the grid index with zero effort that needs effort

#add effort into detection gridIndex
gridEffort$detections$effortArea[effdets[ind]] <- eff
#add appropriate effort into grid cell AREA as needed, then rerun centroids
gridEffort$effort$area[grid_chg] <- eff

centroidAll <- gridCentroid(gridEffort, gpsAll) #assigns the 'pa' column where 0 is no whale, 1 is whale
centroidAll[which(centroidAll$effCells==grid_chg), ]$start <- gridEffort$detections[effdets[ind],5]  #Give the centroid a start and end time, use the detection time
centroidAll[which(centroidAll$effCells==grid_chg), ]$end <- gridEffort$detections[effdets[ind],5]
centroidAll[which(centroidAll$effCells==grid_chg), ]$diff_mins <- 0
```

Resolve effort by using surrounding cell
```{r}
#IF MORE THAN ONE PRESENCE HAS NO EFFORT, DO EACH ONE SEPARATELY, WAAHH
#change values for gridIndex, effortArea, and actualArea in 'gridEffort$detections' while keeping everything else the same

ind = 7 #index of effchg, what detection row needs to be changed? Check ind in gridEffort$detections
CentNew = 4  #represents the row number(s) in GridCentJoin that corresponds with the correct dir, SO S=1, E=2, N=3, W=4
#!! 'newCell' will change depending on the survey and spatial scale #
newCell <- GridCentJoin[CentNew,1] # Select the gridIndex (col 1) for the new cells based on the plot

newEffArea <- gridEffort$effort$area[newCell] # get the effort for each new cell from original grid
newActArea <- actualArea[newCell] # get the new cell's actual area

#This replaces the old cell without effort with info for the new cell in these 3 cols
effchg2 <- effchg[ind,] %>% mutate(gridIndex = newCell, effortArea=newEffArea, actualArea=newActArea)

#remove old effchg and replace with new cell info in the detections list
gridEffort$detections <- gridEffort$detections[-69,] #no effort det to remove
gridEffort$detections <- rbind(gridEffort$detections, effchg2) #bind new cell info with rest of grid detections

#original rows from 'effdets' index should be missing with the new cells at the bottom 
 View(gridEffort$detections)
 
```

#4. Save new grid data
!!! Check that all whale cells have Effort > 0
```{r}
#1706: Excluding 1706_41 peak B to deal with stupid duplicate problem found below.
gridEffort$detections <- gridEffort$detections[-62,] 

saveRDS(gridEffort, here::here( paste0('output/gridEffort/', gridsize, ' km-', loctype), paste0('gridEffort', surveynum, '_', gridsize, 'km_', loctype2, '.rda') ))
```

#5. Run centroids again
```{r}
centroidAll <- gridCentroid(gridEffort, gpsAll)
```


#Check for Duplicate Presences in Centroids
##1. Check that all centroids with presences have effort as expected.

At this point, all presences AND absences should have effort > 0.
Presences can be filtered using detections$gridIndex 

- Now duplicates may occur due including both peaks of the localized acoustic encounters. This is checked using 'dupTRUE'. This does not mean that ALL localized encs occur in the same grid cell, only some do. 
- The 'pa' for centroids with 2 or more different encounters will be changed accordingly after checking they occur within the same month since env data are usually monthly avgs.

```{r}
# find duplicates in centroids
if ((nrow(gridIndex) == length(gridEffort$detections$ID)) == TRUE){
  print('Each detection occurs in its own grid cell')
} else {
  print('There are detections that occur in the same grid cell. Which ones?')
  gridTabl <- gridIndex[which(gridIndex[2] > 1),]
  detInd <- as.integer(gridEffort$detections$gridIndex)
  dupInd <- detInd[duplicated(detInd)] #get value of duplicated grid cell(s)
  gridTabl # table showing grid cells containing > 1 encounter
}

# check that all detections (pa > 0) have effort > 0
all_prz <- dplyr::filter(centroidAll, pa>0) # pa should equal 1 at this point
ifelse( all_prz$EffArea > 0, 'TRUE', 'FALSE' ) #Check that all cells have effort 

#Get detection info for all that have effort, this will exclude any peaks without effort, but includes duplicates too
all_dets <- subset(gridEffort$detections, gridIndex %in% all_prz$effCells)
  
  dupRow <- subset(all_prz, effCells %in% dupInd)      #subset the duplicated rows in CENTROIDS, should be pa=1, which needs to be changed
  dupDets <- subset(gridEffort$detections, gridIndex %in% dupInd)  #subset of duplicated detections. 
  
  # pull duplicate detections that ARE NOT peaks A & B, this will give detections that are truly different.
  dupTRUE <- subset(dupDets, peak == 'trk')
  
  # CHECK they occur in the same month
  month(dupTRUE$UTC)
  
  #update dupInd and dupRow
  dupInd <- dupTRUE$gridIndex[duplicated(dupTRUE$gridIndex)]
  dupRow <- subset(all_prz, effCells %in% dupInd) 

  ## Do this for each 'dupInd' separately if things aren't working right ## YB 4/2022
if(length(dupInd)==0){
  print('Hooray! Each enc is located in its own cell')
} else {
  print('Oopsy! May have >1 enc per cell. Lets sort it out, itll be fun, I swear...Run the following code.' )
}
  centroidAlltmp <- filter(centroidAll, pa != 1)     #remove the rows with single presences
  
  all_single = all_prz[!(all_prz$effCells %in% dupInd), ] #remove duplicated rows based on the effCell value, pa == 1 only

dupRow$pa[na.omit(match(gridTabl$gridIndex, dupRow$effCells))] <- gridTabl$frequency[which(gridTabl$gridIndex %in% dupRow$effCells)]
    # dupRow$pa[ match(gridTabl$gridIndex, dupRow$effCells) ] <- gridTabl$frequency  #same thing as previous line
#further changes to dupRow pa if needed
dupRow$pa[1] <- 2

print(dupRow)                                      #check that the pa column is correct

all_prz2 = rbind(all_single, dupRow)   #add the new row to the rest of the presences

#more checks
which(gridEffort$detections$gridIndex %in% all_prz2$effCells) #length represents total in gridEffort$detections
which(  all_prz2$effCells %in% gridEffort$detections$gridIndex  )  #length represents total in all_prz2


centroidAll <- rbind(centroidAlltmp, all_prz2)       #Should equal the same rows as before since just changing the value in pa




```

##2. Do if centroids with presence is missing effort.

```{r}
### IF DETECTIONS MISSING IN CENTROIDS ###
#Are there any missing detections or mismatches in detections between the grid data and the centroid data? These should all be 0
nrow(all_prz)
# nrow(pmDetsub)
allPrz <- all_prz$effCells
allPrz[!( allPrz %in% gridDets)]

table(gridDets)
table(allPrz)

(notINgrid <- (all_prz$effCells[!all_prz$effCells %in% gridEffort$detections$gridIndex])) 
(notINcent <- (gridEffort$detections$gridIndex[!gridEffort$detections$gridIndex %in% all_prz$effCells]))         


#which detection is missing in the centroid data?
CentInd = which(gridEffort$detections$gridIndex == notINcent) #outputs index
(CentMiss = gridEffort$detections[CentInd,])
# Do more if things don't match, but they should and all should be well.
# (id_not <- which(!(gridIndex$Var1 %in% all_prz$effCells == TRUE))) #find the detections that don't match up
# (not_det <- gridEffort$detections[id_not,]) #this is hopefully empty. These are the detections NOT IN the centroids

```


##3. More Checks
Some cells with presences were passed more than once. These need to have the correct times associated with them to retrieve the correct environmental data. This chunk finds the detections that occur > 1, picks the time that is within a day of the detection UTC, and assigns it to that detection.
THEN, it takes the other cells that don't match the detection UTC and assign it to pa = 0, so it's still counted but not as a presence, pa = 1. 
The same process is done for the absences, where any effort spanning more than 2-3 days are identified and included as separate entries.
```{r}
#### PRESENCE DATA ####
# `dbl_prz` and `dbl_det` should have the same number of rows
#Need to know which time (start or end) is associated with a detection if pa = 1 and diff_mins > 1440 (mins in a day)
dbl_prz <- dplyr::filter(centroidAll, abs(diff_mins) > 1440 & pa > 0)

#In cells with PRESENCE, which detections occur in cells that were passed > 1 time?
dbl_idx <- which(gridEffort$detections$gridIndex %in% dbl_prz$effCells == TRUE)  #could include duplicates
dbl_det <- gridEffort$detections[dbl_idx,]
dbl_det <- gridEffort$detections[dbl_idx,] %>% arrange(gridIndex) #detections in order of gridIndex to match dbl_prz

if ((nrow(dbl_prz) != nrow(dbl_det))==TRUE){ #should be true. IF IT'S FALSE, PROBABLY DUE TO >1 whale in the same cell
#remove duplicated gridIndexes
dbl_det = dbl_det[!duplicated(dbl_det[ , c("gridIndex")]),]
nrow(dbl_prz) == nrow(dbl_det)
} else {
  print('Same nrow')
}
## For some cells with pa = 1, the start and end times are not within a day of the detection time for the same location (cell centroid). The detection times are compared to the start and end times. Those times that differ by less than a day are used. 
idt1 = which(abs(difftime(dbl_det$UTC, dbl_prz$start, units = "mins")) < 1440) #find rows where time difference btwn detection and start is less than a day
idt2 = which(abs(difftime(dbl_det$UTC, dbl_prz$end, units = "mins")) < 1440)   #find rows where time difference btwn detection and end is less than a day
idt3 = which(abs(difftime(dbl_det$UTC, dbl_prz$start, units = 'mins')) > 1440 & abs(difftime(dbl_det$UTC, dbl_prz$end, units = 'mins')) > 1440) #find rows where time difference btwn detection and both start and end is more than a day

gr8end <- dplyr::select(dbl_prz[idt1,], `UTC`= end, effCells, EffArea, diff_mins, lon, lat, pa) #df for end times that occur on diff day than detection
gr8start <- dplyr::select(dbl_prz[idt2,], `UTC`= start, effCells, EffArea, diff_mins, lon, lat, pa) #df for start times that occur on diff day than detection

# Take start and end times when both occur on different day than the detection and create a row for them in a df
gr8both1 <- dplyr::select(dbl_prz[idt3,], `UTC`= end, effCells, EffArea, diff_mins, lon, lat, pa) 
gr8both2 <- dplyr::select(dbl_prz[idt3,], `UTC`= start, effCells, EffArea, diff_mins, lon, lat, pa)
GR8all <- rbind(gr8end, gr8start, gr8both1, gr8both2)
GR8all$pa <- 0 #these are not presences because they are the times that occur on a different day than the detection

# take the actual detection times and match them with the centroid lat lon. 
dbl_det1 <- data.frame(dplyr::select(dbl_det[idt1,], UTC), dplyr::select(dbl_prz[idt1,], effCells, EffArea, diff_mins, lon, lat, pa))
dbl_det2 <- data.frame(dplyr::select(dbl_det[idt2,], UTC), dplyr::select(dbl_prz[idt2,], effCells, EffArea, diff_mins, lon, lat, pa))
dbl_det3 <- data.frame(dplyr::select(dbl_det[idt3,], UTC), dplyr::select(dbl_prz[idt3,], effCells, EffArea, diff_mins, lon, lat, pa))
DBL_prz <- rbind(dbl_det1, dbl_det2, dbl_det3)   ## create df of centroids for all presences with >1 time






#### ABSENCE DATA ####
#In cells with ABSENCE, manipulate to have both times in one column
dbl_abz <- filter(centroidAll, abs(diff_mins) > 3200 & pa == 0) #double effort within 2-3 days? Threshold based on grid data
dbl_start <- dplyr::select(dbl_abz, `UTC`=start)#first time
dbl_end   <- dplyr::select(dbl_abz, `UTC`=end)  #second time
dbl_else <- dplyr::select(dbl_abz, -start, -end)#the rest of the data
dbl_startdf <- data.frame(dbl_start, dbl_else)   #combine first time with the rest of the data
dbl_enddf <- data.frame(dbl_end, dbl_else)       #combine second time with the rest of the data
DBL_abz <- rbind(dbl_startdf, dbl_enddf, GR8all) ## create df of centroids for all absences that had double effort


#Configure cells with single effort to have same format as double effort to combine them
sgl_abz <- dplyr::filter(centroidAll, abs(diff_mins) <= 3200 & pa == 0) # all times are within 2-3 days, which is fine bc datasets are monthly
SGL_abz <- dplyr::select(sgl_abz, `UTC` = start, everything(), -end)

sgl_prz <- dplyr::filter(centroidAll, abs(diff_mins) <= 1440 & pa > 0)
SGL_prz <- dplyr::select(sgl_prz, `UTC` = start, everything(), -end)


####COMBINE ALL!!! ####

ALLcentroids <- rbind(DBL_prz, SGL_prz, DBL_abz, SGL_abz)
ALLcentroids$survey <- surveynum
```
##CHECK FOR NA
```{r}

sum(is.na(ALLcentroids))     # Total NA for presences
(colsum <- colSums(is.na(ALLcentroids))) # Total NA per column
ALLcentroids<- ALLcentroids[complete.cases(ALLcentroids), ]
```
##Save
```{r}
length(unique(pmDetsub$ID)) #number of detections
nrow(dplyr::filter(ALLcentroids, pa > 0)) #should equal number of detections
(gridEffort$detections$gridIndex %in% ALLcentroids$effCells)

# print('Do the number of encounters match the number of centroids with encounters?')
if (nrow(pmDetsub) == nrow(dplyr::filter(ALLcentroids, pa == 1))){
  print('Yes, all encounters match to a centroid. YOU ARE A GENIUS. Now save that shit.')
 saveRDS(ALLcentroids, here::here( paste0('output/centroids/', gridsize, ' km-', loctype) , paste0('centroids', surveynum, '_', gridsize, 'km_', loctype2, '.rda')))
  print('Shit is saved')
} else {
  print('No, but are there > 1 encs in a cell? Check that and it should be all good.')   
  saveRDS(ALLcentroids, here::here(paste0('output/centroids/', gridsize, ' km-', loctype), paste0('centroids', surveynum, '_', gridsize, 'km_', loctype2, '.rda')))
  print('Shit is saved')
}
```


#Plot for Sanity Check
```{r}
# Theres also a plotty function, but it will look pretty messy for all your data. Trackline is (roughly) in blue,
# detections are red dots, amount of effort is shaded in gray, but for a big grid you wont really be able to see the shading
# plotGridResult(gridEffort)

#non-function form of plotting to plot one detection or other data points
x <- gridEffortOLD
actualArea <- as.numeric(st_area(x$grid))
coveragePct <- round(x$effort$area / actualArea, 2)
coveragePct <- ifelse(coveragePct > 1, 1, coveragePct)

pdf(here::here('figures', paste0('Grid', surveynum, '_', gridsize, 'km_', loctype, '_COMB_fig3_orrange.pdf')), width=11,height=8)

plot(x$grid, col = gray(1 - coveragePct, alpha = .9), lwd = 0.2)  #grid
points(x=x$gps$Longitude, y=x$gps$Latitude, col='darkorange', pch=16, cex = 0.05)  #trackline
points(centroidAll$lon, centroidAll$lat, col='red', pch = 20, cex=0.08, lwd=0.1)  #centroids

# points(x=centroidAll$lon, y=centroidAll$lat, col='red', pch = 16, cex=0.5 ) #point for A121.S44
#203.4318	21.36744	three times through here?

# points(x=dbl_eff2$lon, y=dbl_eff2$lat, col='red', pch = 0, cex=0.2, lwd = 0.2 ) 
# points(x=pmDetsub$Longitude, y=pmDetsub$Latitude, col='green', pch=1, cex=0.3, lwd=0.3) #Longitude is trackline location
# text(pmDetsub$Longitude, pmDetsub$Latitude,col='magenta', labels=pmDetsub$acid, cex=0.3, lwd = 0.3 )
# points(x=187.6831, y=26.60179	, col='darkturquoise', pch = 15, cex=0.5, lwd = 0.5 ) #square
# points(x=187.5513, y=26.66397	, col='deeppink3', pch = 17, cex=0.5, lwd = 0.5 ) #triangle
# points(effchg$Longitude, effchg$Latitude, col='orange', pch=3, cex=0.05)

dev.off()
# points(effchg$lon2, effchg$lat, col='red', pch = 0, cex=0.2, lwd = 0.2 )
# points(all_prz$lon, all_prz$lat,col='goldenrod', pch = 16, cex=0.2, lwd = 0.2 )


# plot(pmDetsub$lon, pmDetsub$Latitude,col='magenta', pch = 17, cex=0.2, lwd = 0.2 )


# points(201.8474, 22.30299,col='green', pch = 16, cex=0.2, lwd = 0.2 )
# points(202.0274, 22.30299,col='darkmagenta', pch = 16, cex=0.2, lwd = 0.2 )
# points(201.7574, 22.30299,col='deepskyblue', pch = 16, cex=0.2, lwd = 0.2 )
# points(201.9824, 22.34799,col='darkorange', pch = 0, cex=0.2, lwd = 0.2 ) #15053 centroid ->CHOOSE THIS ONE
# points(202.0724, 22.34799,col='darkmagenta', pch = 0, cex=0.2, lwd = 0.2 )#15054 centroid

# points(effchg$Longitude, effchg$Latitude, col='darkred', pch = 16, cex=0.2, lwd = 0.2 ) #dot, 1303_A20
# points(effchg$lon_trk2, effchg$lat_trk, col='darkorange2', pch = 16, cex=0.2, lwd = 0.2 ) #trackline location
# points(surroundCells$lon[1], surroundCells$lat[1], col='aquamarine2', pch = 16, cex=0.2, lwd = 0.2 )
# points(surroundCells$lon[2], surroundCells$lat[2], col='coral2', pch = 16, cex=0.2, lwd = 0.2 )
# points(surroundCells$lon[3], surroundCells$lat[3], col='darkorchid2', pch = 16, cex=0.2, lwd = 0.2 )
# 
# 
# 
# #From straightpath_yb
# ggplot(str8_1303, aes(x=Longitude, y=Latitude)) +
#             geom_point(aes(group=timeGroup, col=straight), size = 1.3) +
#             scale_color_manual(values = c('red', 'darkgreen')) +
#             ggtitle(paste0('nSmall=', nSmall, ", nLarge=", nLarge)) +
#   geom_point(x=-167.28, y=24.287, col='magenta')

```

## Centroid Consolidate!
```{r}
CombCents = NULL
# AcOnlyEnv = NULL
# pa1Check = NULL
for (s in c(1641, 1303, 1604, 1705, 1706)){
  combi <- readRDS(here::here( paste0('output/centroids/', gridsize, ' km-', loctype, '/',  #new AcOnly centroids
                                        'centroids', s, '_', gridsize, 'km_', loctype2, '.rda') ))

  CombCents = rbind(CombCents, combi)  
  
}
saveRDS(CombCents, here::here( paste0('output/centroids/', gridsize, ' km-', loctype, '/',  #new Combined centroids
                                        'centroids_AllSurveys', '_', gridsize, 'km_', loctype2, '.rda') ))

```
