==> code/REVISIONS/8_GetSSTchlAWavPow_2022.Rmd <==
---
title: "8_GetSSTchlAWavPow_v4"
author: "Yvonne Barkley"
date: "9/13/2020"
output: html_document
editor_options: 
  chunk_output_type: console
---
devtools::install_github('TaikiSan21/PAMmisc')
```{r setup, include=FALSE}
library(PAMmisc)
library(PAMpal)

library(here)
```

```{r}
#### Which Survey? ####
surveynum = 'AllSurveys'
gridsize = 16
loctype = 'Combined'
loctype2 = 'Comb'

# centroidData <- sw
centroidData <- readRDS(here::here( paste0('output/envData/', gridsize, ' km-', loctype, '/', 
                                           surveynum, '_', gridsize, 'km_', loctype2, '_sw.rda') ))


colnames(centroidData)[5] <- "Longitude"
colnames(centroidData)[6] <- "Latitude"
# [c(1,2)] <- c("good", "better")
```
#2-3 of 4. Run centroids with no buffer, getting data from single location of the centroid and saving it

#SST
```{r}

envDatSST = 'erdMH1sstdmday'
httpSST = 'https://coastwatch.pfeg.noaa.gov/erddap/'
edinfoSST <- erddapToEdinfo(envDatSST, baseurl = httpSST, chooseVars = FALSE)
edinfoSST <- varSelect(edinfoSST, select = c(TRUE))




```

```{r}
envType = 'sst'
matchedData <- matchEnvData(centroidData, nc=edinfoSST, fileName = here::here('output/envData', paste0(gridsize, ' km-', loctype, '/', surveynum, "_", gridsize, 'km_', loctype2, '_', envType, ".nc")))

edinfo = edinfoSST
matchedDataSST <- matchedData

# SSTsd
matchedDataSSTsd2 <- matchEnvData(centroidData, nc=edinfoSST, buffer=c(.08, .08, 0), FUN=sd)
matchedData<-matchedDataSSTsd2                          
edinfo = 

matchedDataSSTsd2[matchedDataSSTsd2[,9]=='NA'] <- NA 
which(is.na(matchedDataSSTsd2[,9]))
```

#CHLA
```{r}
 
envDatCHLA = 'hawaii_soest_69ee_7d0d_74e6'

httpCHLA = 'http://apdrc.soest.hawaii.edu/erddap/'
edinfoCHLA <- erddapToEdinfo(envDatCHLA, baseurl = httpCHLA, chooseVars = FALSE)
edinfoCHLA <- varSelect(edinfoCHLA, select = c(TRUE))
```

```{r}
envType = 'chla'
matchedDataCHLA <- matchEnvData(centroidData, nc=edinfoCHLA) 
                            fileName = here::here('output/envData', paste0(gridsize, ' km-', loctype, '/',
                                                                           surveynum, "_", gridsize, 'km_',
                                                                           loctype2, '_', envType, ".nc")))
edinfo = edinfoCHLA
matchedDataCHLA[matchedDataCHLA[,9]=='NA'] <- NA                  #put in NaNs to variable column
matchedData<- matchedDataCHLA
```

#WP
```{r}

envDatWP = 'ww3_global' #old ID: 'hawaii_soest_98bb_253a_eb1c'
httpWP = 'https://pae-paha.pacioos.hawaii.edu/erddap/' #'http://apdrc.soest.hawaii.edu/erddap/'

edinfoWP <- erddapToEdinfo(envDatWP, baseurl = httpWP)
```


```{r}
envType = 'wp'

matchedDataWP <- matchEnvData(centroidData, nc=edinfoWP)

matchedDataWP$wavepow <-  ((1024*(9.8^2))/(64*pi))*matchedDataWP$Thgt_mean^2*matchedDataWP$Tper_mean

edinfo <- edinfoWP
saveRDS(matchedDataWP, paste0('output/envData/', gridsize, ' km-', loctype, '/', surveynum, '_', gridsize, 'km_',loctype2, '_wavepow.rda') )

# rm(list = ls()[grep("EnvData", ls())])  #use to remove variables starting with the same name

```


#SSH / SSHsd / EKE

```{r}
edinfoSSH <-paste0(here::here('data'), '/SSH.nc')
envType = 'temp600'
matchedDataSSH <- matchEnvData(centroidData, nc=edinfoSSH)
matchedData2 <- matchEnvData(centroidData, nc=edinfoSSH, buffer(.04, .04, 0), FUN=c(mean))


edinfo = edinfoSSH
matchedData3[matchedData3[,11]=='NA'] <- NA                  #put in NaNs to variable column


#SD of SSH
matchedDataSSHsd <-matchEnvData(centroidData,nc=edinfoSSH, buffer=c(.04, .04, 0), FUN=sd)
matchedDataSSHsd[matchedDataSSHsd[,11]=='NA'] <- NA                  #put in NaNs to variable column

saveRDS(matchedDataSSH, paste0('output/envData/', gridsize, ' km-', loctype, '/', surveynum, '_', gridsize, 'km_',loctype2, '_', envType, '.rda') )


#After finishing SSH, compute EKE
matchedDataSSH$eke = 0.5*(matchedDataSSH$uo_mean^2 + matchedDataSSH$vo_mean^2)
saveRDS(matchedDataSSH, paste0('output/envData/', gridsize, ' km-', loctype, '/', surveynum, '_', gridsize, 'km_',loctype2, '_', envType, '.rda') )

```

#Temp at 600 m
```{r}
envType = 'temp600'
envDatTemp600 = 'hawaii_soest_d346_28ac_fccf'
httpTemp600 = 'http://apdrc.soest.hawaii.edu/erddap/'

edinfoTemp600 <- erddapToEdinfo(envDatTemp600, baseurl = httpTemp600)


matchedDataTemp600 <- matchEnvData(centroidData, nc=edinfoTemp600, depth = 584) 
                            # fileName = here::here('output/envData', paste0(gridsize, ' km-', loctype, '/',
                                                                           # surveynum, "_", gridsize, 'km_',
                                                                           # loctype2, '_', envType, ".nc"))

matchedDataTemp600$temp600 <- matchedDataTemp600$potdsl_mean-273.15

saveRDS(matchedDataTemp600, paste0('output/envData/', gridsize, ' km-', loctype, '/', surveynum, '_', gridsize, 'km_',loctype2, '_', envType, '.rda') )

```


#RUN MEGA LOOP                                                                                                    
```{r}

matchedData <- matchedDataSSHsd[, -c(9,10)]
#### MEGA IFELSE ####
# Series of ifelse statements to check for NAs and fill them using the 'buffer' argument with 1km steps to get data
if (nrow(matchedData[is.na(matchedData[,9]),]) > 0){ #if NAs exist

matchedNA <- matchedData[is.na(matchedData[,9]),]         #make dataframe of NA rows
#Subset complete cases  
matchedDataComplete <- matchedData[complete.cases(matchedData[,9]),]  #make dataframe of complete rows in variable column
    
    #Run match function again even if there are no NAs, it will otherwise go to the 'else' statement.
matchedData2 <- matchEnvData(matchedNA[,1:8], nc=edinfo, buffer = c(gridsize/100+0.001,gridsize/100+0.001,0)) #use buffer of 1km to get data outside the grid size in either direction on same day
# matchedData2 <- matchEnvData(matchedNA[,1:8], nc=edinfo, buffer = c(gridsize/100+0.01,gridsize/100+0.01,0)) #use buffer of 1km to get data outside the grid size in either direction on same day
    matchedData2[matchedData2[,9]=='NA'] <- NA

        if (nrow(matchedData2[is.na(matchedData2[,9]),]) > 0){ #if using 0.11 buffer results in NAs
        
        matchedNA2 <- matchedData2[is.na(matchedData2[,9]),]  #make second dataframe of NA rows
        matchedDataComplete2 <- matchedData2[complete.cases(matchedData2[,9]),]  #make dataframe of complete rows in variable column
            
          #Run match function again
        matchedData3 <- matchEnvData(matchedNA2[,1:8], nc=edinfo, buffer = c(gridsize/100+0.002,gridsize/100+0.002,0)) #use buffer to get data outside the grid size in either direction on same day
        matchedData3[matchedData3[,9]=='NA'] <- NA
        
                if (nrow(matchedData3[is.na(matchedData3[,9]),]) > 0){ #if using 0.12 buffer still produces NAs
                matchedNA3 <- matchedData3[is.na(matchedData3[,9]),]  #make third dataframe of remaining NA rows
                matchedDataComplete3 <- matchedData3[complete.cases(matchedData3[,9]),]  #make dataframe of complete rows 
                
                  #Run match function again
                  matchedData4 <- matchEnvData(matchedNA3[,1:8], nc=edinfo, buffer = c(gridsize/100+0.003,gridsize/100+0.003,0)) #use buffer to get data outside the grid size in either direction on same day
                  matchedData4[matchedData4[,9]=='NA'] <- NA
                  
                          if (nrow(matchedData4[is.na(matchedData4[,9]),]) > 0){
                          matchedNA4 <- matchedData4[is.na(matchedData4[,9]),]  #make third dataframe of remaining NA rows
                          matchedDataComplete4 <- matchedData4[complete.cases(matchedData4[,9]),]  #make dataframe of complete rows 
                    
                            #Run match function again
                            matchedData5 <- matchEnvData(matchedNA4[,1:8], nc=edinfo, buffer = c(gridsize/100+0.1,gridsize/100+0.1,0)) #use buffer to get data outside the grid size in either direction on same day
                            matchedData5[matchedData5[,9]=='NA'] <- NA
                            
                                    if (nrow(matchedData5[is.na(matchedData5[,9]),]) > 0){ #if NA's still remain
                                      print('Reevaluate how to fill NAs')
                                      matchedNA5 <- matchedData5[is.na(matchedData5[,9]),]  #make third dataframe of remaining NA rows
                                      matchedDataComplete5 <- matchedData5[complete.cases(matchedData5[,9]),]  #make dataframe of complete rows 
                                    } else {
                                    print(paste0('No NAs, finally! Used buffer=', gridsize/100+0.1, ' for final NAs'))
                                    matchedDataFinal <- rbind(matchedDataComplete, matchedDataComplete2, matchedDataComplete3, matchedDataComplete4, matchedData5)
                                    print('The matched env data has the same number of rows as the centroid dataset') 
                                    nrow(matchedDataFinal) == nrow(centroidData)
                                    }
                  
                    
                        } else { #if no NAs in matchedData4 
                        print(paste0('No NAs, finally! Used buffer=', gridsize/100+0.003))
                        matchedDataFinal <- rbind(matchedDataComplete, matchedDataComplete2, matchedDataComplete3, matchedData4)
                        print('The matched env data has the same number of rows as the centroid dataset') 
                        nrow(matchedDataFinal) == nrow(centroidData)
                        }
                  
                } else { #if no NAs in matchedData3 
                  print(paste0('No NAs, finally! Used buffer=', gridsize/100+0.002))
                  matchedDataFinal <- rbind(matchedDataComplete, matchedDataComplete2, matchedData3)
                  print('The matched env data has the same number of rows as the centroid dataset') 
                  nrow(matchedDataFinal) == nrow(centroidData)
                }
                
        } else { #if no NAs in matchedData2
          print(paste0('No NAs, yay! Used buffer=', gridsize/100+0.001))
            matchedDataFinal <- rbind(matchedDataComplete, matchedData2)
          print('The matched env data has the same number of rows as the centroid dataset') 
            nrow(matchedDataFinal) == nrow(centroidData)
        }


        
} else { #if no NAs in matchedData, ie: no NAs during the first data grab
      print('No NAs, yay! All centroids filled on first try. No buffer necessary.')
      matchedDataFinal <- matchedData
      print('The matched env data has the same number of rows as the centroid dataset') 
     
    
}

#4 of 4 SAVE!
EnvData_tmp=assign(paste0('EnvData_', envType), matchedDataFinal) 
saveRDS(EnvData_tmp, paste0('output/envData/', gridsize, ' km-', loctype, '/', surveynum, '_', gridsize, 'km_',loctype2, '_', envType, '.rda') )

#calculate wavepower if needed
if (envType == 'wp'){
  EnvData_tmp$wavepow = ((1024*(9.8^2))/(64*pi))*EnvData_tmp$Thgt_mean^2*EnvData_tmp$Tper_mean
  # EnvData_tmp$wavepow = ((1024*(9.8^2))/(64*pi))*EnvData_tmp$htsgwsfc_mean^2*EnvData_tmp$perpwsfc_mean
  EnvData_tmp <- EnvData_tmp[, c(1:8, 24, 9:23)] #move wavepow to col 9 to match other datasets using other script
  saveRDS(EnvData_tmp, paste0('output/envData/', gridsize, ' km-', loctype, '/', surveynum, '_', gridsize, 'km_',loctype2, '_wavepow.rda') )
  
}
nrow(matchedDataFinal) == nrow(centroidData)
nrow(matchedDataFinal) == nrow(centroidSub)

envType
```

IF SUBSETTING TOTAL DATASETS for WAVE POWER to WORK, COMBINE ALL SUBSETS HERE
```{r}
#(code adapted from EffortCheck.r)

surveynum = 'AllSurveys'
AcOnlyWavepow = NULL
for (s in c(1641, 1303, 1604, 1705, 1706)){
  aconlyi <- readRDS(here::here( paste0('output/envData/', gridsize, ' km-', loctype, '/',  #new AcOnly centroids
                                       s, '_', gridsize, 'km_', loctype2, '_wavepow.rda') ))


  AcOnlyWavepow = rbind(AcOnlyWavepow, aconlyi)  

}

  saveRDS(AcOnlyWavepow, paste0('output/envData/', gridsize, ' km-', loctype, '/AllSurveys_', gridsize, 'km_',loctype2, '_wavepow.rda') )

```


# EnvData_SST = matchedDataFinal
# saveRDS(EnvData_SST, here::here('output/envData', paste0(surveynum,'_10km_', envType, '_',loctype2, '.rda')))
# 
# EnvData_CHL = matchedDataFinal
# saveRDS(EnvData_CHL, here::here('output/envData', paste0(surveynum, '_10km_', envType, '_',loctype2, '.rda')))
