==> code/REVISIONS/10_PmSDM_Foraging_2022.Rmd <==
---
title: "Summary of SDMs for Foraging and Non-Foraging Groups"
author: "Yvonne Barkley"
date: "10/18/2020"
output:
 pdf_document:
    
    latex_engine: xelatex
    extra_dependencies: ["flafter"]
editor_options: 
  chunk_output_type: console
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, comment = '', fig.width =8, fig.height = 6, message=FALSE, tidy.opts=list(width.cutoff=60))
```

### This document summarizes the results of comparing sperm whale distribution models for:
  * FORAGING and NON-FORAGING WHALE groups.

Both models were built using the combined data set, which included sightings with concurrent acoutic encounters and acoustic encounters without an associated sighting.


 * **Foraging Groups**    
     * Groups with regular clicks or creaks identified with an inter-click interval between 0.5 - 1 s 
     * n = 104 foraging encounters
 
 * **Non-Foraging Groups**    
     * Groups without regular clicks or creaks. These models included groups with codas or slow clicks present.
     * n = 76 non-foraging encounters

$~$

### Environmental Variables  

All models incorporated biologically relevant environmental predictors that represented bathymetric features or oceanographic processes that act as proxies for mechanisms driving sperm whale prey distribution. The correlation between environmental variables was tested prior to model fitting. All environmental variables included in the models (listed in the table below) resulted in correlation coefficients < |0.6|.
Chlorophyll, EKE and wave power were log-transformed to reduce the skew and minimize the leverage of outlying data points during model fitting.

The spatial plots of the dynamic variables also provide an overview of the gradients within the environmental data.


```{r echo=FALSE}
## image ####
require(here)
require(knitr)
include_graphics(here('figures/EnvDataTable3.png'))
```
```{r echo=FALSE, fig.cap='Spatial plots of all dynamic environmental variables tested in the models.', fig.width=4, fig.height=6}
## image ####
require(here)
require(knitr)
include_graphics(here('figures/SwMap_AllDyn2.png'))
```


$~$


### Model Construction
A 25-km spatial resolution was used to compute the unit of analysis (25 x 25 grid cell) for the number of sperm whale groups per unit. Preliminary models also tested a 10-km spatial resolution, but increasing the spatial resolution did not result in any significant changes to the strength of correlations between encounter rate and environmental variables. 

Model data sets were divided into train and test data sets using a 70/30 split. GAMs were fitted using a negative binomial with a log-link function as this is an appropriate distribution for over-dispersed data sets consisting of relatively sparse count data with large numbers of zeros. All smoothers were fit using thin-plate regression splines (the default basis). Each simple smoother was limited to 3 degrees of freedom (k=3) to reduce overfitting parameters per recommendations from other studies building similar types of cetaceans distribution models. A spatial smoother ( s(Lon, Lat, k=10) ) was also included in all models to account for spatial autocorrelation within the data. The number of knots was increased to 10 based on model diagnostics to account for the complexity of the spatial smoother and minimize overfitting the data. The log of the effort was also included as an offset to account for the variation in search effort per grid cell. 

* Model variable selection was performed in two ways to select variables using:
  * Forward selection starting with a spatial smoother since there is spatial structure within the data set. Each environmental variables was added to the spatial smoother model and assessed using AIC, explained deviance, and whether it was statistically significant (p < 0.05).
  * Backward selection starting with a full model using all variables. Non-significant variables were removed and the model was refitted until all remaining variables were significant.
  * Selecting the final variables included comparing models with AIC during both selection procedures. Then, the best models were compared to select the most parsimoniuous set of variables.


```{r, echo=F, eval=FALSE, message=FALSE}
library(tidyverse)
library(mgcv)
library(corrplot)
library(geoR)
library(tidymv)
library(here)
library(lubridate)
library(viridis)
library(dplyr)
splitdf <- function(dataframe, seed=NULL) {
    if (!is.null(seed)) set.seed(seed)
    index <- 1:nrow(dataframe)
    trainindex <- sample(index, trunc(length(index)*0.8))
    trainset <- dataframe[trainindex, ]
    testset <- dataframe[-trainindex, ]
    list(trainset=trainset,testset=testset)
}

#randomly select As and Bs for loc encs: https://stackoverflow.com/questions/10010220/random-row-selection-in-r
randomRows = function(df,n){
  return(df[sample(nrow(df),n),])
}
```

```{r echo=FALSE}
#Values for COMBINED MODELS
survey = 'AllSurveys'
gridsize = 16
loctype = 'Combined'
loctype2 = 'Comb'
#data has logged vars and temporal variables
## SAVED from line 251 in 10_PmSDM_AcousticCombined.rmd
# PmCombo <- readRDS(here::here( paste0('output/models/', loctype, '/data/', 'CompletePm_', gridsize, 'km_', loctype2, '_EDIT.rda') ))
# PmCombo2 <- readRDS(here::here( paste0('output/models/', loctype, '/data/', 'CompletePm_', gridsize, 'km_', loctype2, '_scaled.rda') ))

PmCombo <- readRDS( here::here(paste0('data/Pm_Combined_2022_v2.rda') ) )


# rm(list = ls()[grep("ssh", ls())]) 
```

# YB Revisits 
2/7/22 Hello old friend
A few encounters need to have the click codes changed as a result of closer inspection during ACCURATE click analysis. The changes will be made in PmCombo prior to splitting the train and test data and any foraging/non-foraging analysis.
Changes are determined by the yellow-highlighted 'click_code' cells in the ACCURATE Project Sheet.

### 2022: Redo Click Codes -> saved as 'v2'
```{r}
PmCombo$click_code[PmCombo$ID == "1303_42_21"] <- "cdckrs"  #(was ckrs, added coda)
PmCombo$click_code[PmCombo$ID == "1604_30_999"] <- "rs"     #(was s)
PmCombo$click_code[PmCombo$ID == "1604_67_999"] <- "rs"     #(was s)
PmCombo$click_code[PmCombo$ID == "1705_251_999"] <- "rs"    #(was s)
PmCombo$click_code[PmCombo$ID == "1706_2_999"] <- "rs"      #(was r)
PmCombo$click_code[PmCombo$ID == "1706_41_999"] <- "rs"     #(was s)
# PmCombo$click_code[PmCombo$ID == "1706_135_999"] <- "ckrs" #not included as separate encounter, was combined with 1706A49 in same cell, which is also ckrs
PmCombo$click_code[PmCombo$ID == "1303_71_999"] <- "rs"     #(was s)

saveRDS(PmCombo, here::here(paste0('data/Pm_Combined_2022_v2.rda') ))

colnames(PmCombo)[c(18:27, 47:49)] <- c('SST', 'Chla', 'Temp584m', 'WavePower', 'Depth', 'DistShore', 'SSH', 'SSHsd', 'EKE', 'DistSeamt', 'log.Chla', 'log.EKE', 'log.WavePower')

PmComboPres <- filter(PmCombo, pa >0)
plot(PmComboPres$Longitude, PmComboPres$DistShore)
min(PmComboPres$DistShore)
max(PmComboPres$DistShore)


PmComboLoc <- filter(PmCombo, loc ==1 )
plot(PmComboLoc$ID, PmComboLoc$DistShore)
min(PmComboLoc$DistShore)
max(PmComboLoc$DistShore)

```




### Saving new train/test data in 2022 Revisions folders
"C:/Users/yvers/Documents/CHP 3/SpermWhales/output/models/Combined/data/2022_revisions/Train_25km_Comb_revise.rda"
```{r eval=FALSE, echo=FALSE}
## COMBO: REDO TRAIN TEST for COMBO MODELS ####
require(dplyr)
splitdf <- function(dataframe, seed=NULL) {
    if (!is.null(seed)) set.seed(seed)
    index <- 1:nrow(dataframe)
    trainindex <- sample(index, trunc(length(index)*0.7))
    trainset <- dataframe[trainindex, ]
    testset <- dataframe[-trainindex, ]
    list(trainset=trainset,testset=testset)
}

trainComb = NULL
testComb = NULL
seed = 2

for (s in c(1641, 1303, 1604, 1705, 1706)){
   
 trSub <- filter(PmCombo, survey == s)

 #subset for presences and split 70/30
 pres1 <- filter(trSub, pa > 0 )# & loc == 1) #include all presence data/acoustic encounters
 listPres  <- splitdf(pres1, seed) #output is list for train and test

 #subset for absences and split 70/30
 abs0  <- filter(trSub, pa == 0 )
 listAbs <- splitdf(abs0, seed)  #output is list for train and test

 #combine train data for presence and absence
 trainAll <- rbind( listPres$trainset, listAbs$trainset )
 
 #combine test data for presence and absence
 testAll  <- rbind( listPres$testset,  listAbs$testset  )
   
trainComb = rbind( trainComb, trainAll )
testComb  = rbind( testComb,  testAll )

# trainAcOnly$log.effort <- log(trainAcOnly$EffArea)
# testAcOnly$log.effort <- log(testAcOnly$EffArea)
}
saveRDS(trainComb, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Train_', gridsize, 'km_', loctype2, '_revise.rda')  ))
saveRDS(testComb, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Test_', gridsize, 'km_', loctype2, '_revise.rda')  ))

# nrow(dplyr::filter(trainAcOnly, trainAcOnly$pa >0))
# nrow(dplyr::filter(testAcOnly, testAcOnly$pa >0))
```

```{r echo=FALSE}
## COMBO: TRAINING AND TEST DATA #### 

trainComb <- readRDS(here::here(  paste0('output/models/',loctype, '/data/Train_',   gridsize, 'km_', loctype2, '_revise.rda')  ))
testComb <- readRDS(here::here(  paste0('output/models/',loctype, '/data/Test_',     gridsize, 'km_', loctype2, '_revise.rda')  ))

# trainComb <- readRDS(here::here(  paste0('output/models/',loctype, '/data/Train_',   gridsize, 'km_', loctype2, '_Comb3.rda')  ))
# testComb <- readRDS(here::here(  paste0('output/models/',loctype, '/data/Test_',     gridsize, 'km_', loctype2, '_Comb3.rda')  ))

```




## Foraging and Non-Foraging Group Models

Foraging models incorporated a subset of the sperm whale encounters from the Combined Models (sighted with acoustics, and acoustics only) that only included encounters with **regular clicks** and **creaks** to indicate foraging.


```{r eval=FALSE,echo=FALSE}
## Forage and Social Group Data Subset ####
#1. Denote forage or not. 
PmFor1 <- filter(PmCombo, grepl('r|k', click_code))   # the 'k' implies ck for creaks because c also for codas
PmFor0 <- filter(PmCombo, !grepl('r|k', click_code))
PmFor1$forage <- 1  #forage, includes peaks A & B   
PmFor0$forage <- 0  #non-forage
PmForage <- dplyr::bind_rows(PmFor1, PmFor0)


#2. Denote non-foraging
PmPositive <- filter( PmCombo, pa > 0 )
PmAbsent <- filter( PmCombo, pa == 0 )

PmNonFor1 <- filter(PmPositive, !grepl('r|k', click_code))  #subset all encs without r or ck in click_code
PmNonFor0 <- filter(PmPositive, grepl('r|k', click_code))

#designate whether the encounters included codas (1) or not (0)
PmNonFor1$nonfor <- 1
PmNonFor0$nonfor <- 0
PmAbsent$nonfor <- 0

PmNonFor <- rbind(PmNonFor1,PmNonFor0, PmAbsent)


# PmNonFor2 <- PmNonFor[, c(1,2,4,54,62)]

nrow(PmNonFor1)  # represent all non-foraging 
nrow(PmNonFor0)  # represent all foraging encounters, which are now 0
```

# FORAGE 2022: REDO Train Test ####
Using a 80 / 20 split for training and test data sets
```{r eval=FALSE, echo=FALSE}

Pm_uniq <- plyr::ddply(PmForage, .(ID), randomRows, 1) #unique IDs
Pm_uniq2 <- plyr::ddply(Pm_uniq, .(effCells), randomRows, 1) #unique effCells to solve pa==2 repeats
Pm_abs  <- filter( PmForage, pa == 0 )

PmForage2 <- rbind(Pm_uniq2, Pm_abs)

# colnames(PmForage2)[c(18:27, 47:49)] <- c('SST', 'Chla', 'Temp584m', 'WavePower', 'Depth', 'DistLand', 'SSH', 'SSHsd', 'EKE', 'DistSmt', 'log.Chla', 'log.EKE', 'log.WavePower')

# 
# test <-filter(PmForage, pa > 0 & forage == 1)
# CountForage <- test[, c(1:10, 23,41, 51)]


trainForg = NULL
testForg = NULL
# preschk = NULL
seed = 1 #124a #200e

for (s in c(1641, 1303, 1604, 1705, 1706)){
   require(dplyr)
 trSub <- filter(PmForage2, survey == s)

 #subset for presences and split 70/30
 pres1 <- filter(trSub, pa > 0 & forage == 1)# & loc == 1) #include all presence data/acoustic encounters
 listPres  <- splitdf(pres1, seed) #output is list for train and test
# preschk = rbind(preschk, pres1)
 #subset for absences and split 80/20
 abs0  <- filter(trSub, pa == 0 )
 listAbs <- splitdf(abs0, seed)  #output is list for train and test

 #combine train data for presence and absence
 trainAll <- rbind( listPres$trainset, listAbs$trainset )
 
 #combine test data for presence and absence
 testAll  <- rbind( listPres$testset,  listAbs$testset  )
   
trainForg = rbind( trainForg, trainAll )
testForg  = rbind( testForg,  testAll )

# trainAcOnly$log.effort <- log(trainAcOnly$EffArea)
# testAcOnly$log.effort <- log(testAcOnly$EffArea)
}

#keep in mind these include the log variables too
saveRDS(trainForg, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Train_', gridsize, 'km_', loctype2, '_Forage80a.rda')  ))
saveRDS(testForg, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Test_', gridsize, 'km_', loctype2, '_Forage80a.rda')  ))

# nrow(dplyr::filter(trainAcOnly, trainAcOnly$pa >0))
# nrow(dplyr::filter(testAcOnly, testAcOnly$pa >0))
```

```{r eval=FALSE, echo=FALSE}
# Non-Foraging: REDO Train Test ####
Pm_uniq <- plyr::ddply(PmNonFor, .(ID), randomRows, 1) #unique IDs
Pm_uniq2 <- plyr::ddply(Pm_uniq, .(effCells), randomRows, 1) #unique effCells to solve pa==2 repeats

Pm_abs  <- filter( PmNonFor, pa == 0 )

PmNonFor <- rbind(Pm_uniq2, Pm_abs)


trainNonFor = NULL
testNonFor = NULL
# preschk = NULL
seed = 1

for (s in c(1641, 1303, 1604, 1705, 1706)){
   require(dplyr)
 trSub <- filter(PmNonFor, survey == s)

 #subset for presences and split 80/20
 pres1 <- filter(trSub, pa > 0 & nonfor == 1)# & loc == 1) #include all presence data/acoustic encounters
 listPres  <- splitdf(pres1, seed) #output is list for train and test
# preschk = rbind(preschk, pres1)
 #subset for absences and split 70/30
 abs0  <- filter(trSub, pa == 0 )
 listAbs <- splitdf(abs0, seed)  #output is list for train and test

 #combine train data for presence and absence
 trainAll <- rbind( listPres$trainset, listAbs$trainset )
 
 #combine test data for presence and absence
 testAll  <- rbind( listPres$testset,  listAbs$testset  )
   
trainNonFor = rbind( trainNonFor, trainAll )
testNonFor  = rbind( testNonFor,  testAll )

# trainAcOnly$log.effort <- log(trainAcOnly$EffArea)
# testAcOnly$log.effort <- log(testAcOnly$EffArea)
}
saveRDS(trainNonFor, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Train_', gridsize, 'km_', loctype2, '_NonFor80a.rda')  ))
saveRDS(testNonFor, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Test_', gridsize, 'km_', loctype2, '_NonFor80a.rda')  ))

# nrow(dplyr::filter(trainAcOnly, trainAcOnly$pa >0))
# nrow(dplyr::filter(testAcOnly, testAcOnly$pa >0))
```
### Load Train and Test data
```{r echo=FALSE}
## FORAGE/NON-FORAGE: LOAD TRAIN and TEST ####
# trainHunt <- trainForg
# testHunt <- testForg

#seed 1
trainHunt <- readRDS(here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Train_',   gridsize, 'km_', loctype2, '_Forage80a.rda')  ))
testHunt <- readRDS(here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Test_',     gridsize, 'km_', loctype2, '_Forage80a.rda')  ))


# trainHunt25km <- readRDS(here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Train_', '25km_', loctype2, '_HuntRevise2.rda')  ))

colnames(trainHunt)[18] <- 'SST'
colnames(trainHunt)[19] <- 'log.Chla'
colnames(trainHunt)[20] <- 'Temp584m'
colnames(trainHunt)[24] <- 'SSH'
colnames(trainHunt)[25] <- 'SSHsd'
colnames(testHunt)[18] <- 'SST'
colnames(testHunt)[19] <- 'log.Chla'
colnames(testHunt)[20] <- 'Temp584m'
colnames(testHunt)[24] <- 'SSH'
colnames(testHunt)[25] <- 'SSHsd'
#Removing 2 chla outliers, see notes in Data Analysis checklist, Oct 8, 2020. Doh, kept them in...IDK
# trainHunt <- filter(trainHunt, log.chla < 5)

trainHunt <- filter(trainHunt, SSHsd < 0.03 )
testHunt <- filter(testHunt, SSHsd < 0.03)
#seed 1

trainNonFor <- readRDS(here::here(paste0('output/models/',loctype, '/data/2022_revisions/Train_',   gridsize, 'km_', loctype2, '_NonFor80a.rda')  ))
testNonFor <- readRDS(here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Test_',     gridsize, 'km_', loctype2, '_NonFor80a.rda')  ))

colnames(trainNonFor)[18] <- 'SST'
colnames(trainNonFor)[19] <- 'Depth'
colnames(trainNonFor)[20] <- 'log.Chla'
colnames(trainNonFor)[32] <- 'Temp584m'
colnames(trainNonFor)[38] <- 'SSH'
colnames(trainNonFor)[39] <- 'SSHsd'

colnames(testNonFor)[30] <- 'SST'
colnames(testNonFor)[34] <- 'Depth'
colnames(testNonFor)[57] <- 'log.Chla'
colnames(testNonFor)[32] <- 'Temp584m'
colnames(testNonFor)[38] <- 'SSH'
colnames(testNonFor)[39] <- 'SSHsd'

trainNonFor <- filter(trainNonFor, SSHsd < 0.03 )
testNonFor <- filter(testNonFor, SSHsd < 0.03)


###TOTALS for Whale Encounters ####
PmForage_total <- subset(PmForage, pa >0 & forage ==1)
PmForage_totalb <- select(PmForage_total, pa, ID, forage, loc)
sum(PmForage_total$pa) # since pa can be > 1
sum(subset(PmForage_totalb, pa >0 & forage ==1 & loc==1)$pa) #localized forage
sum(subset(PmForage_totalb, pa >0 & forage ==1 & loc==0)$pa) #trackline and sighted forage



PmNonFor_total <- subset(PmNonFor, pa >0 & nonfor ==1)
sum(PmNonFor_total$pa)
PmNonFor_totalb <- select(PmNonFor_total, pa, ID, nonfor, loc)

```


#### 1) Foraging Model  

  * gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(log.Chla, k=3) + s(SSHsd, k=3)
  * The spatial smoother accounts for the spatial variation not explained by the other predictors in the model. There are obviously some other factors at play that are responsible for the distribution patterns of the foraging whales, but the spatial smoother at least helps to explain the extra variation. A model with only the spatial smoother resulted in an explained deviance of ~12%, but had a higher AIC value relative to the best model (689.7 vs 677.4, respectively).
  * The effect of the Temp at 584m is in the form of a 'U distribution' with a minimum at ~6.8°C. Higher encounter rates were predicted at the minimum and maximum temperatures at 584 m, which are associated with more whale encounters in the cooler region north of the MHI and warmer temperatures north of the NWHI region. The depth of these temperatures reflect the potential temperatures of sperm whale prey habitat. Whether the temperatures directly affect the prey distribution is unclear, but this is one of the more effective variables for explaining the occurrence of sperm whales compared to the other variables.
  * Log-transformed Chl-a was also an important variable despite the counterintuitive negative relationship with encounter rate. There are 2 outliers of higher Chl-a concentrations. Models were tested excluding these outliers and Chl-a remained statistically significant with the negative relationship. 
  * SSHsd resulted in a similar U-distribution as Temp at 584m. SSHsd predicted higher encounter rates at the minimum and maximum SSHsd values (~0 and 0.025 m), with the lowest encounter rates occurring within the middle of the range at ~0.012 m. The lower SSHsd corresponds with the high encounter rate in the MHI region. The higher SSHsd corresponds with the high encounter rate in the NWHI region north and west of Laysan Atoll. This may indicate more mesoscale eddy activity in this area, which is somewhat validated by the spatial plot showing slightly more values of increased eddy kinetic energy in that region. The reasons or processes for higher number of foraging whales in the MHI are less obvious.
    


# 2022 REVISED MODELS ~~~~~~~

## FORAGE 2022

Tried included quarter as both a smoother and a factor -> results were the same.
```{r}

# didn't include quarter since sampling is biased
modForageTW  <- gam(pa ~  s(Longitude, Latitude, k=12) + s(Depth, k=3) + s(DistShore, k=3)  + s(DistSeamt, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.EKE, k=3) + s(log.WavePower, k=3) + s(year, k=3) + s(array, k=3) + offset(log.effort), data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTW) 

#train80a - USED in paper
modForageTW2  <- gam(pa ~  s(Longitude, Latitude, k=12) +  s(Temp584m, k=3) + s(log.Chla, k=3) + offset(log.effort) , data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTW2) 
concurvity(modForageTW2, full=FALSE)

#train80d
modForageTW3  <- gam(pa ~  s(Longitude, Latitude, k=15) + s(DistLand, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + offset(log.effort) , data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTW3) 

#train70c
modForageTW5  <- gam(pa ~  s(Longitude, Latitude, k=12) + s(SSH, k=3) + s(Temp584m, k=3) + offset(log.effort) , data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTW5)   #6.94%

#all foraging
modForageTW4  <- gam(pa ~  s(Longitude, Latitude, k=15) + s(SST, k=3) + s(log.Chla, k=3) + s(SSHsd, k=3)+ s(log.WavePower, k=3) + offset(log.effort) , data = PmForage2, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTW4) 


par(mar=c(4,4,3,3),mfrow = c(2,2))
gam.check(modForageTW4)

saveRDS(modForageTW2, here::here(paste0('output/models/Combined/data/2022_revisions/GAM_best_forage.rda') ) )

```

## FORAGE-NO s(x,y)
```{r}
modForageTWb  <- gam(pa ~ s(Depth, k=3) + s(DistShore, k=3)  + s(DistSeamt, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.EKE, k=3) + s(log.WavePower, k=3) + offset(log.effort) +  s(year, k=3) , data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTWb) 

modForageTWc  <- gam(pa ~  s(Temp584m, k=3) + s(SSH, k=3) + offset(log.effort) , data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTWc) 

modForageNBb  <- gam(pa ~ s(Depth, k=3) + s(DistShore, k=3)  + s(DistSeamt, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.EKE, k=3) + s(log.WavePower, k=3) + offset(log.effort) +  s(year, k=3) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForageNBb) 

modForageNBc  <- gam(pa ~ s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForageNBc)

```


#Forage Plots
### Foraging Effects Plots
```{r echo=FALSE, fig.cap = "Foraging Model: Partial effects plots for the spatial contour, Temperature at 584 m, log.Chla, and SSHsd." , fig.width=8, fig.height=9}

modForage <-  modForageTW2

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/Foraging_EffectsPlot_16km.png", width = 8, height = 3, res = 300, units = "in") # 8x6"

par(mar=c(4,4.5,1,1),mfrow = c(1,2))

# plot(modForage, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='SSHsd (m)',seWithMean = TRUE, ylim = c(-5,5),cex.axis = 1.2, cex.lab = 1.5 )

# plot(modForage, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab=expression("SST (°C)"), seWithMean = TRUE, ylim = c(-1.5,2),cex.axis = 1.2, cex.lab = 1.5 )

plot(modForage, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='Temp at 584 m (°C)' , seWithMean = TRUE, ylim = c(-2,2),cex.axis = 1.2, cex.lab = 1.5)#, shift = coef(modForageBest)[1],	trans	= function(x)	exp(x))

plot(modForage, select=3, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab=expression("log(Chl-a) (mg/m"^3*")"), seWithMean = TRUE, ylim = c(-3,2),cex.axis = 1.2, cex.lab = 1.5 )

dev.off()
# mtext(paste0("Foraging Whales"), side=3, line=1, outer=TRUE, cex=1, font=1)

```

### Foraging 2D smoother plot

** Save using the R Export and choose 1200 x 800 to get smaller black dots.
```{r eval=FALSE, echo=FALSE, fig.cap = "Foraging Model: Contour plot of spatial smoother. Purple dots represent acoustically detected encounters. Black dots are all data points with effort included in the model.", fig.width=6, fig.height=4}

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/Foraging_Smoother2Dplot_16kmB.png", width = 32, height = 16, res = 300, units = "cm") # 8x6"

# pdf(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/Foraging_Smoother2Dplot_16kmB.pdf", width = 7, height = 5)

par(mfrow = c(1,1),mar=c(4.5,5,2,1))

plot(modForage, select = 1, scheme = 2, lwd = 2, cex.axis = 1.2, cex.lab = 1.5, main='',
     contour.col = 'black', hcolors=magma(50) , pch=20, cex=.7) #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 16, col='deepskyblue', #mediumslateblue #000DFF
       lwd=1.25, cex=1.2)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.001, col='white')

dev.off()
```

Use for color legend for 2D spatial smoother
```{r}
# vis.gam(modForageBest, view = c("Longitude", "Latitude"), plot.type = "contour", contour.col = "blue", color = "heat")
library(mgcViz)
getIt <- getViz(modForage)
getIt2 <- plot(sm(getIt, 1))

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/1_Foraging_Smoother2Dlegend_16km.png", width = 9, height = 5, res = 300, units = "in") # 8x6"

plot(sm(getIt, 1)) + l_fitRaster() + l_fitContour() +
  scale_fill_gradientn(colours = magma(50)) +l_points(shape=16, size=0.3, alpha=0.25)

dev.off()
# plot(modForageBest, allTerms = T) #plot all terms



```


#NON-FORAGE 2022

```{r}
modNonForage   <- gam(pa ~  s(Longitude, Latitude, k=15) + s(Depth, k=3) + s(DistShore, k=3)  + s(DistSeamt, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.EKE, k=3) + s(log.WavePower, k=3) + offset(log.effort) +  s(year, k=3) , data = trainNonFor, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modNonForage) 
par(mar=c(4,4,3,3),mfrow = c(2,2))
gam.check(modNonForage)

#train nonforage 80a - AVED and USED in paper
modNonForage   <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SSHsd, k=3) + offset(log.effort) , data = trainNonFor, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modNonForage)
concurvity(modNonForage, full=FALSE)


saveRDS(modNonForage, here::here(paste0('output/models/Combined/data/2022_revisions/GAM_best_nonforage.rda') ) )



modNonForageTW <- gam(pa ~  s(Longitude, Latitude, k=12) + s(bath_m, k=3) + s(dist, k=3)  + s(d2smt, k=3) +  s(sst_mean, k=3) + s(log.chla, k=3) + s(temp600, k=3) + s(ssh, k=3) + s(sshsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + s(year, k=3) + offset(log.effort) , data = trainNonFor, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modNonForageTW) 
gam.check(modNonForageTW)


modNonForageTW2 <- gam(pa ~  s(Longitude, Latitude, k=12) + s(log.chla, k=3) + s(sshsd, k=3) + s(log.wp, k=3) + offset(log.effort) , data = trainNonFor, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modNonForageTW2) 
gam.check(modNonForageTW2)




```
## NONFORAGE-NO s(x,y)
```{r}
modNonForageTWb  <- gam(pa ~ s(Depth, k=3) + s(DistShore, k=3)  + s(DistSeamt, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.EKE, k=3) + s(log.WavePower, k=3) + offset(log.effort) +  s(year, k=3) , data = trainNonFor, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modNonForageTWb) 

modNonForageTWc  <- gam(pa ~ s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + offset(log.effort) , data = trainNonFor, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modNonForageTWc) 

```

### NonForaging Effects Plots
```{r echo=FALSE, fig.cap = "Foraging Model: Partial effects plots for the spatial contour, Temperature at 584 m, log.Chla, and SSHsd." , fig.width=8, fig.height=9}
 
png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/NonForaging_EffectsPlot_16km_b.png", width = 4, height = 3, res = 300, units = "in") # 8x6"

par(mar=c(4,4.5,1,1),mfrow = c(1,1))

# plot(modNonForageTW2, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab=expression("log(Chl-a) (mg/m"^3*")"), seWithMean = TRUE, ylim = c(-3,2),cex.axis = 1.2, cex.lab = 1.5 )

plot(modNonForage, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='SSHsd (m)',seWithMean = TRUE, ylim = c(-5,1),cex.axis = 1.2, cex.lab = 1.5 )

# plot(modNonForageTW2, select=4, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='Wave Power (kW/m)',seWithMean = TRUE, ylim = c(-2,2),cex.axis = 1.2, cex.lab = 1.5 )

dev.off()
# mtext(paste0("Foraging Whales"), side=3, line=1, outer=TRUE, cex=1, font=1)

```

### NonForaging 2D smoother plot
** Save using the R Export and choose 1200 x 800 to get smaller black dots.

```{r eval=FALSE, echo=FALSE, fig.cap = "Foraging Model: Contour plot of spatial smoother. Purple dots represent acoustically detected encounters. Black dots are all data points with effort included in the model.", fig.width=6, fig.height=4}

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/NonForaging_Smoother2Dplot_16km.png", width = 9, height = 5, res = 300, units = "in") # 8x6"

par(mfrow = c(1,1),mar=c(4.5,5,2,1))

plot(modNonForage, select = 1, scheme = 2, lwd = 2, cex.axis = 1.2, cex.lab = 1.5, main='',
     contour.col = 'black', hcolors=magma(50) , pch=20, cex=.7) #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainNonFor, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 16, col='deepskyblue', #mediumslateblue
       lwd=1.25, cex=1.2)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.001, col = 'white')

dev.off()
```

Use for color legend for 2D spatial smoother
```{r}
# vis.gam(modForageBest, view = c("Longitude", "Latitude"), plot.type = "contour", contour.col = "blue", color = "heat")
# library(mgcViz)
getIt <- getViz(modNonForage)
getIt2 <- plot(sm(getIt, 1))

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/1_NonForaging_Smoother2Dlegend_16km.png", width = 9, height = 5, res = 300, units = "in") # 8x6"

plot(sm(getIt, 1)) + l_fitRaster() + l_fitContour() +
  scale_fill_gradientn(colours = magma(50)) +l_points(shape=16, size=0.3, alpha=0.25)

dev.off()
# plot(modForageBest, allTerms = T) #plot all terms



```



# Combined All Encounters

```{r}
# Combined Train Test ####


Pm_uniq <- plyr::ddply(PmCombo, .(ID), randomRows, 1) #unique

Pm_uniq2 <- plyr::ddply(Pm_uniq, .(effCells), randomRows, 1) #unique effCells to solve pa==2 repeats
Pm_abs  <- dplyr::filter( PmCombo, pa == 0 )

PmCombo <- rbind(Pm_uniq2, Pm_abs)


require(dplyr)
trainAllEncs = NULL
testAllEncs = NULL
# preschk = NULL
seed = 777

for (s in c(1641, 1303, 1604, 1705, 1706)){
   
 trSub <- dplyr::filter(PmCombo, survey == s)

 #subset for presences and split 70/30
 pres1 <- dplyr::filter(trSub, pa > 0)# & loc == 1) #include all presence data/acoustic encounters
 listPres  <- splitdf(pres1, seed) #output is list for train and test
# preschk = rbind(preschk, pres1)
 #subset for absences and split 70/30
 abs0  <- filter(trSub, pa == 0 )
 listAbs <- splitdf(abs0, seed)  #output is list for train and test

 #combine train data for presence and absence
 trainAll <- rbind( listPres$trainset, listAbs$trainset )
 
 #combine test data for presence and absence
 testAll  <- rbind( listPres$testset,  listAbs$testset  )
   
trainAllEncs = rbind( trainAllEncs, trainAll )
testAllEncs  = rbind( testAllEncs,  testAll )

# trainAcOnly$log.effort <- log(trainAcOnly$EffArea)
# testAcOnly$log.effort <- log(testAcOnly$EffArea)
}
saveRDS(trainAllEncs, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Train_', gridsize, 'km_', loctype2, '_AllEncs80.rda')  ))
saveRDS(testAllEncs, here::here(  paste0('output/models/',loctype, '/data/2022_revisions/Test_', gridsize, 'km_', loctype2, '_AllEncs20.rda')  ))

colnames(trainAllEncs)[20] <- 'Temp584m'
colnames(trainAllEncs)[25] <- 'SSHsd'


```


```{r}
modAllEncs <- gam(pa ~  s(Longitude, Latitude, k=12) + s(bath_m, k=3) + s(dist, k=3)  + s(d2smt, k=3) +  s(sst_mean, k=3) + s(log.chla, k=3) + s(temp600, k=3) + s(ssh, k=3) + s(sshsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + s(year, k=3) + offset(log.effort) , data = trainAllEncs, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modAllEncs) 

modAllEncs2 <- gam(pa ~  s(Longitude, Latitude, k=12) +  s(Temp584m, k=3) + s(SSHsd, k=3) + s(log.eke, k=3) + offset(log.effort) , data = trainAllEncs, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modAllEncs2) 

modAllEncs3 <- gam(pa ~  s(Longitude, Latitude, k=10) +  s(temp600, k=3) + s(sshsd, k=3) + s(log.eke, k=3) + offset(log.effort) , data = trainAllEncs, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modAllEncs3) 

saveRDS(modAllEncs2, here::here(paste0('output/models/Combined/data/2022_revisions/GAM_best_AllEncs.rda') ) )


```


### All Encounters Effects Plots
```{r echo=FALSE, fig.cap = "Foraging Model: Partial effects plots for the spatial contour, Temperature at 584 m, log.Chla, and SSHsd." , fig.width=8, fig.height=9}
 
png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/AllEncs_EffectsPlot_16km.png", width = 10, height = 3, res = 300, units = "in") # 8x6"

par(mar=c(4,4.5,1,1),mfrow = c(1,3))

plot(modAllEncs2, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='Temp at 584 m (°C)' , seWithMean = TRUE, ylim = c(-1,1.2),cex.axis = 1.2, cex.lab = 1.5)#, shift = coef(modForageBest)[1],	trans	= function(x)	exp(x))

plot(modAllEncs2, select=3, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='SSHsd (m)',seWithMean = TRUE, ylim = c(-5,1),cex.axis = 1.2, cex.lab = 1.5 )

plot(modAllEncs2, select=4, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='log.EKE (m2/s2)',seWithMean = TRUE, ylim = c(-1,3),cex.axis = 1.2, cex.lab = 1.5 )

dev.off()
# mtext(paste0("Foraging Whales"), side=3, line=1, outer=TRUE, cex=1, font=1)

```

### All Encounters 2D smoother plot
```{r eval=FALSE, echo=FALSE, fig.cap = "Foraging Model: Contour plot of spatial smoother. Purple dots represent acoustically detected encounters. Black dots are all data points with effort included in the model.", fig.width=6, fig.height=4}

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/NonForaging_Smoother2Dplot_16km.png", width = 9, height = 5, res = 300, units = "in") # 8x6"

par(mfrow = c(1,1),mar=c(4.5,5,2,1))

plot(modAllEncs2, select = 1, scheme = 2, lwd = 2, cex.axis = 1.2, cex.lab = 1.5, main='',
     contour.col = 'black', hcolors=magma(50) , pch=20, cex=.7) #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainNonFor, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 16, col='#000DFF', #mediumslateblue
       lwd=1.25, cex=1.2)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.1)

dev.off()
```

Use for color legend for 2D spatial smoother
```{r}
# vis.gam(modForageBest, view = c("Longitude", "Latitude"), plot.type = "contour", contour.col = "blue", color = "heat")
# library(mgcViz)
getIt <- getViz(modAllEncs2)
getIt2 <- plot(sm(getIt, 1))

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/1_AllEncs_Smoother2Dlegend_16km.png", width = 9, height = 5, res = 300, units = "in") # 8x6"

plot(sm(getIt, 1)) + l_fitRaster() + l_fitContour() +
  scale_fill_gradientn(colours = magma(50)) +l_points(shape=16, size=0.3, alpha=0.25)

dev.off()
# plot(modForageBest, allTerms = T) #plot all terms



```

#Predictions 2022
```{r echo=FALSE}

require(magrittr)
require(dplyr)

### Foraging Model Predictions ####
twTrainForage <- trainHunt %>% mutate(resid = resid(modForageTW2), predict = predict(modForageTW2))
predTrain1 <- predict.gam(modForageTW2, type = 'response', se.fit=TRUE)  #calculate MSE for these to compare with test set. If they're super different, speaks to the genrality of the model
twTrainForage$fit <- predTrain1$fit
twTrainForage$se.fit <- predTrain1$se.fit
#using scale of 0,1,2 makes this hard to interpret
twForageMSE <- mean((twTrainForage$pa - twTrainForage$fit)^2)  #MSE


# mean(abs((nbTrainFinal$pa - nbTrainFinal$fit)))  #Mean absolute error
## Calculate MSE AFTER transforming the predictions back to the same scale as the observed data

nbPred1 <- predict.gam(modForageTW2, newdata = testHunt, type = 'response', se.fit = TRUE)
nbTestFinal1 <- data.frame(testHunt, fit = nbPred1$fit, se.fit=nbPred1$se.fit)
nbMSEtest1 <- mean((nbTestFinal1$pa - nbTestFinal1$fit)^2) #MSE


### Non-Foraging Model Predictions ####
#pulling the prediction and residual data from the model
twTrainNonForage <- trainNonFor %>% mutate(resid = resid(modNonForage), predict = predict(modNonForage))
predTrainLL1 <- predict.gam(modNonForage, type = 'response')  #calculate MSE for these to compare with test set. If they're super different, speaks to the genrality of the model
twTrainNonForage$fit <- predTrainLL1

#using scale of 0,1,2 makes this hard to interpret
twNonForageMSE <- mean((twTrainNonForage$pa - twTrainNonForage$fit)^2)  #MSE
# mean(abs((nbTrainFinal$pa - nbTrainFinal$fit)))  #Mean absolute error
## Calculate MSE AFTER transforming the predictions back to the same scale as the observed data

nbPredLL1 <- predict.gam(modNonForage, newdata = testNonFor, type = 'response', se.fit = TRUE)
nbTestLL1 <- data.frame(testNonFor, fit = nbPredLL1$fit, se.fit=nbPredLL1$se.fit)
nbMSEtestLL1 <- mean((nbTestLL1$pa - nbTestLL1$fit)^2) #MSE

# mean(abs((testFinal$pa - testFinal$fit))) #Mean absolute error

# AIC
AICforage <- AIC(modForageTW2)
AICNonForage <- AIC(modNonForage)

# Explained Deviance
twForage_expDev = round(((modForageTW2$null.deviance-modForageTW2$deviance)/modForageTW2$null.deviance)*100, 2)
twNonForage_expDev = round(((modNonForage$null.deviance-modNonForage$deviance)/modNonForage$null.deviance)*100, 2)

# Summary table
# 
table2 = matrix(NA, nrow = 2, ncol = 6)
colnames(table2) = c("Models", "ExpDev", "AIC", "MSEtrain", "MSEtest", "Predictor Variables")

### HARD CODED SOME VALUES BC THEY WERE BEING A LITTLE BITCH
table2[1,] <- c("Foraging", paste0(twForage_expDev, '%'), round(AICforage,2), round(twForageMSE,3), round(nbMSEtest1,3), 
              paste0('Temp at 584m, log.Chla, LON:LAT'))
table2[2,] <- c("Non-Foraging", paste0(twNonForage_expDev, '%'), round(AICNonForage,2), round(twNonForageMSE,3),
               round(nbMSEtestLL1,3), paste0('SSHsd, LON:LAT'))
require(knitr)
# kable(table, caption = "Summary Results for Foraging and Non-Foraging Models")
kable(table2, caption = "Summary Results for Foraging and Non-Foraging Models")
```


#OLD 2020
    
### OLD MODELS ###
```{r echo=FALSE, eval=FALSE}
## FORAGE MODEL ####
# trainHuntb <- dplyr::filter(trainHunt, chla.r < 0.3 ) #even removing outliers, chla still signif

## START HERE WITH 2D SMOOTH
# modForage2Donly <- gam(pa ~  s(Longitude, Latitude, k=10) + s(depth.r, k=3) + s(distland.r, k=3)  + s(distseamt.r, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + offset(log.effort) + s(year, k=3) + qtr2 , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
# summary(modForage) 

modForage       <- gam(pa ~  s(Longitude, Latitude, k=10) + s(depth.r, k=3) + s(distland.r, k=3)  + s(distseamt.r, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + + s(year, k=3) + qtr + offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForage) 

modForage2 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(log.Chla, k=3) + s(SSHsd, k=3) + s(SST, k=3) +offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML") #remove qtr and that makes sst non-sig
summary(modForage2) 

modForage3 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + s(SSHsd, k=3) +s(log.Chla, k=3) + offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForage3) 

modForage4 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(SSHsd, k=3) + offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForage4) 

############ TW #######

modForageTW <- gam(pa ~  s(Longitude, Latitude, k=10) + s(depth.r, k=3) + s(distland.r, k=3)  + s(distseamt.r, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + offset(log.effort) + s(year, k=3) + qtr , data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTW) 

modForageTW2 <- gam(pa ~  s(Longitude, Latitude, k=10)  + s(Temp584m, k=3) + s(SSHsd, k=3) + s(qtr, k=3) , data = trainHunt, family = tw, link = 'log', select = TRUE, method = "REML")
summary(modForageTW2) 

par(mar=c(4,4,3,2),mfrow = c(2,2))
gam.check(modForage2)
gam.check(modForageTW2)

########################
modForg2D2 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSHsd, k=3) + s(qtr, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForg2D2) 

modForg2D3 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + s(Temp584m, k=3) + s(qtr, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForg2D3)


# Using all Forage data
modForageAll <- gam(pa ~  s(Longitude, Latitude, k=10) + s(depth.r, k=3) + s(distland.r, k=3)  + s(distseamt.r, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + offset(log.effort) , data = PmFor, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForageAll) 
modForageAll <- gam(pa ~  s(Longitude, Latitude, k=10) + s(distseamt.r, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSHsd, k=3) + offset(log.effort) , data = PmFor, family = nb, link = 'log', select = TRUE, method = "REML")


```


```{r echo=FALSE}
require(mgcv)
## OLD- FORAGE MODEL BEST ####
modForageBest <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(log.Chla, k=3) + s(SSHsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modForageBest)

modForageBest2 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) +  s(SSHsd, k=3), offset = log.effort, data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
# summary(modForageBest2)

modForagePlot <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(log.Chla, k=3) + s(SSHsd, k=3), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
# summary(modForagePlot)
#?############ other models #######
# modForageBest <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + s(Temp584m, k=3) + s(log.Chla, k=3) + s(SSHsd, k=3) + offset(log.effort) + s(qtr, k=3) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
# summary(modForageBest)  #SST depends on qtr, so removing both
#?
# 
# 
# modForageBestOLD <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + s(Temp584m, k=3) + s(qtr, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
# summary(modForageBestOLD) #which now has sst as non-sig

 ## log Chla only
# modForageBest2 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(log.Chla, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest2)
#  
##  Temp 584 only
# modForageBest3 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest3)

 ## temp584, sshsd only
#  modForageBest4 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(SSHsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest4)
# 
 ## sshsd only
#  modForageBest5 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SSHsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest5)
# 
# # log chla and sshsd only 
#  modForageBest7 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(log.Chla, k=3) + s(SSHsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest7)
# 
#  modForageBest8 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + s(SSHsd, k=3) + s(Temp584m, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest8) #no chla
#  
# modForageBest9 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + s(SSHsd, k=3) + s(log.Chla, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest9) #no temp584
#  
## Lon:Lat only
# modForageBest10 <- gam(pa ~  s(Longitude, Latitude, k=10) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest10)

#   # temp584, chl-a only 
# modForageBest11 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(log.Chla, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modForageBest11)
# 
# AIC(modForageBest,modForageBest2,modForageBest3,modForageBest4,modForageBest5,modForageBest7,modForage, modForage3, modForageBest10,modForageBest11)

## Lon:Lat only
# modJeff1 <- gam(pa ~  s(Longitude, Latitude, k=10) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modJeff1)
#  
#  modJeff2 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SSH, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modJeff2)
# 
#   modJeff3 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SST, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#  summary(modJeff3)
#  
#    modJeff4 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(SSHsd, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#   summary(modJeff4)
#   
#  modJeff5 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#   summary(modJeff5)
#  modJeff6 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(log.Chla, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#   summary(modJeff6)
# 
#  modJeff7 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(log.eke, k=3) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
#   summary(modJeff7)
#   
#   AIC(modJeff1,modJeff2, modJeff3, modJeff4, modJeff5, modJeff6, modJeff7 )
```
##
Set up plots without Chl-a, modForage4
```{r echo=FALSE, fig.cap = "Foraging Model: Partial effects plots for the spatial contour, Temperature at 584 m and SSHsd." , fig.width=8, fig.height=9}

mlay <- rbind(c(1,1), c(2,3))
# print(mlay)
# layout(mlay)
# layout.show(4)

layout(mlay)
par(mar=c(4,4,1,1))
# par(mar=c(4,4,3,2),mfrow = c(2,2))
# plot(modForageBest)
plot(modForage4, select = 1, scheme = 2, lwd = 2) #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 10, col='mediumslateblue', lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)

plot(modForage4, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='Temp at 584 m (°C)' , seWithMean = TRUE, ylim = c(-2,2))#, shift = coef(modForageBest)[1],	trans	= function(x)	exp(x))

plot(modForage4, select=3, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='SSHsd (m)',seWithMean = TRUE, ylim = c(-2,2) )

mtext(paste0("Foraging Whales"), side=3, line=1, outer=TRUE, cex=1, font=1)

```


Visualize the raw data
```{r echo=FALSE, fig.cap = "Foraging Model: The encounter rate (Groups per cell) plotted separately as a function of Temperature at 584 m, log.Chla, and SSHsd."}

par(mar=c(4,4,1,2),mfrow = c(3,1))

plot(trainHunt$Temp584m, trainHunt$pa, ylab="Groups per cell", xlab="Temp at 584m")
plot(trainHunt$log.Chla, trainHunt$pa, ylab="Groups per cell", xlab="log.Chla")
plot(trainHunt$SSHsd, trainHunt$pa, ylab="Groups per cell", xlab="SSHsd")

```



$~$
#### 2) Non-Foraging Model  

  * gam(pa ~  s(Longitude, Latitude, k=10) + s(Depth, k=3)
  
  * Depth and the spatial smoother were the two variables that best explained the data for the non-foraging groups. The U-shaped relationship of depth predicted the most groups at the deeper and shallower depths. The least groups were predicted for depths between 3000-3500 m. 
  * Non-foraging groups consisted of slow clicks and codas, sometimes both within the same encounter. It's possible that there are different factors driving the distribution of non-foraging whales that were not included in the model. This could be related to reproductive behavior since the non-foraging groups likely include males and social groups of females and jueveniles. 
  * Note: When only the spatial smoother was fitted for the non-foraging model, it performed nearly as well as the model that also included depth. This implies that depth (and the other nonsignificant environmental variables) do not provide much inforation for explaining the distribution of non-foraging groups. 
  
```{r echo=FALSE, eval=FALSE}
### BEST NON-FORAGING MODEL w/ 2D Smoother ####
modNonFor <- gam(pa ~  s(Longitude, Latitude, k=12) + s(Depth, k=3) + s(distland.r, k=3)  + s(distseamt.r, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + offset(log.effort), data = trainNonFor, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modNonFor) 

modNonFor2 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Depth, k=3) + offset(log.effort) , data = trainNonFor, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modNonFor2) 

modNonFor2 <- modNonForBest
saveRDS(modNonForBest,  here::here( paste0('code/REVISIONS/BEST_NonForagingModel.rda') ))

modNonForBest <-  readRDS(here::here( paste0('code/REVISIONS/BEST_NonForagingModel.rda') ))
```

### Non-Foraging Effects Plots
```{r echo=FALSE, eval=FALSE}


png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/NonForaging_EffectsPlot_b.png", width = 4, height = 4, res = 300, units = "in") 

par(mar=c(4,4,1,3),mfrow = c(1,1))
plot(modNonForBest, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='Depth (m)' , seWithMean = TRUE, ylim = c(-2,3) )

dev.off()

```
### Non-Foraging 2D plot
```{r}
png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/NonForaging_EffectsPlot.png", width = 9, height = 5, res = 300, units = "in") # 8x6"

par(mfrow = c(1,1),mar=c(4.5,5,2,1))

plot(modNonForBest, select = 1, scheme = 2, lwd = 2, cex.axis = 1.2, cex.lab = 1.5, main='',
     contour.col = 'black', hcolors=magma(50)) #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainNonFor, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 16, col='#000DFF', #mediumslateblue
       lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)

dev.off()
```

```{r}
#Get the color legend from this plot
getIt_NonFor <- getViz(modNonForBest)
getIt2 <- plot(sm(getIt, 1))

png(file = "C:/Users/yvers/OneDrive/Documents/Publications/2022_FiRS_SDMs/figures/NonForaging_Smoother2Dlegend.png", width = 9, height = 5, res = 300, units = "in") 

plot(sm(getIt_NonFor, 1)) + l_fitRaster() + l_fitContour() +
  scale_fill_gradientn(colours = magma(50)) +l_points(shape=16, size=0.3, alpha=0.25)

dev.off()
```





```{r echo=FALSE, fig.cap = "Non-Foraging Model selected depth and spatial smoother, which was significant despite the straight contour lines. Purple dots represent acoustically detected encounters. Black dots are all data points with effort included in the model.", fig.width=8, fig.height=9}

par(mfrow = c(1,1), mar=c(5,5,3,1))

plot(modNonForBest, select=2, residuals = FALSE, pch = 20, shade = T, scheme = 2, shade.col = 'gray85', all.terms = TRUE, xlab='Depth (m)' , seWithMean = TRUE, cex.axis = 1.2, cex.lab = 1.5 )

plot(modNonForBest, select = 1, scheme = 2, lwd = 2,cex.axis = 1.2, cex.lab = 1.5, main='' ) #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainNonFor, pa >0) #get points for whales to plot

#for 2D smoother plot
points(whales$Longitude, whales$Latitude, pch = 16, col='mediumslateblue', lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)
```
#Non-forage contour plot
```{r}
getIt <- getViz(modNonForBest)
getIt2 <- plot(sm(getIt, 1))

plot(sm(getIt, 1)) + l_fitRaster() + l_fitContour() +
  scale_fill_gradientn(colours = heat.colors(50))

```

```{r echo=FALSE, fig.cap = "Side-by-side comparison of spatial plots for Foraging and Non-Foraging models. Foraging whales represent groups with regular clicks and creaks. Non-foraging whales represent groups with codas and slow clicks present.", fig.width=8, fig.height=9}
par(mfrow = c(2,1), mar=c(4,4,1,2))
## SIDE BY SIDE ####
#Foragers
plot(modForageBest, select = 1, scheme = 2, lwd = 2, main = 'Foragers') #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot
points(whales$Longitude, whales$Latitude, pch = 10, col='mediumslateblue', lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)

# modForage2Donly <-gam(pa ~  s(Longitude, Latitude, k=10) + offset(log.effort), data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
# plot(modForage2Donly, select = 1, scheme = 2, lwd = 2, main = 'Foragers-spatial smooth only')
# summary(modForage2Donly)

## Non-Foragers
plot(modNonForBest, select = 1, scheme = 2, lwd = 2, main = 'Non-Foragers') #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainNonFor, pa >0) #get points for whales to plot
points(whales$Longitude, whales$Latitude, pch = 10, col='mediumslateblue', lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)

# modNonForage2Donly <-gam(pa ~  s(Longitude, Latitude, k=10) + offset(log.effort), data = trainNonFor, family = nb, link = 'log', select = TRUE, method = "REML")
# plot(modNonForage2Donly, select = 1, scheme = 2, lwd = 2, main = 'Foragers-spatial smooth only')
# summary(modNonForage2Donly)

```





$~$



```{r echo=FALSE, fig.cap= 'The higher density areas with more foraging groups are further illustrated with a bimodal distribution of longitudes where the two peaks correspond with the spatial contour plots. The distribution of latitudes are more normally distributed, but there is evidence of peaks between 22-24°N and 26-28°N. Few whales occurred below 20°N.'  }
## Foraging Lat:Lon ####

PmFor = rbind(trainHunt, testHunt)
PmFor2 <- subset(PmFor, pa >0)
saveRDS(PmFor, here::here( paste0('data/Pm_ForagingOnly.rda') ))
# PmFor2$Longitude = ifelse(PmFor2$Longitude > 180, PmFor2$Longitude-360, PmFor2$Longitude)
# lon = ifelse(sw$lon>180, sw$lon-360, sw$lon)
par(mfrow = c(2,2), mar=c(4,4,1,2))

# Foraging
hist(PmFor2$Latitude, breaks = 5, main= '', xlab='Latitude of Foraging Groups')
plot(PmFor$Latitude, PmFor$pa, ylab="Foraging Groups per cell", xlab="Latitude (All Data)")

hist(PmFor2$Longitude, breaks = 10, main = '', xlab='Longitude of Foraging Groups')
plot(PmFor$Longitude, PmFor$pa, ylab="Foraging Groups per cell", xlab="Longitude (All Data)")
```

```{r echo=FALSE, fig.cap= 'Non-foraging groups are distributed evenly across latitudes with more occurring in the western portion of the study area. Few whales occurred below 20°N. '}
par(mfrow = c(2,2), mar=c(4,4,1,2))

## NonForaging Lat:Lon ####
PmNonFor = rbind(trainNonFor, testNonFor)
PmNonFor2 <- subset(PmNonFor, pa >0)
saveRDS(PmNonFor, here::here( paste0('data/Pm_NonForagingOnly.rda') ))

# Non-Foraging
hist(PmNonFor2$Latitude, breaks = 5, main= '', xlab='Latitude of Non-Foraging Groups')
plot(PmNonFor$Latitude, PmNonFor$pa, ylab="Non-Foraging Groups per cell", xlab="Latitude (All Data)")

hist(PmNonFor2$Longitude, breaks = 10, main = '', xlab='Longitude of Non-Foraging Groups')
plot(PmNonFor$Longitude, PmNonFor$pa, ylab="Non-Foraging Groups per cell", xlab="Longitude (All Data)")

```

```{r eval=TRUE, echo=FALSE, fig.cap='Contour plot of temperature at 584 m.' }


modTemp584 <- gam(Temp584m ~  s(Longitude, Latitude, k=100) , data = PmFor, select = TRUE, method = "REML") 
modTemp584 <- gam(Temp584m ~  s(Longitude, Latitude, k=10) , data = PmFor, select = TRUE, method = "REML") 
modTemp584 <- gam(Temp584m ~  s(Longitude, Latitude, k=50) , data = PmFor, select = TRUE, method = "REML") 

summary(modTemp584) 

par(mfrow = c(1,1))

plot(modTemp584, select = 1, scheme = 2, lwd = 2, main = 'Temp at 584m ~ LAT:LON, k=10') #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot
points(whales$Longitude, whales$Latitude, pch = 10, col='mediumslateblue', lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)

```
```{r echo=FALSE, eval=FALSE}
## FAIL: NO NEED-contour plot Temp584m ####
require(akima)
require(plotly)
library(pracma)

# Looking at temp at depth for Foraging groups

ctour =ggplot(PmFor, aes(Longitude, Latitude, z= Temp584m)) 
ctour +  geom_density_2d_filled()
ctour + geom_density_2d()
ctour +geom_contour()




a = interp(PmFor$Longitude,PmFor$Latitude,PmFor$Temp584m, duplicate = "mean")

plot_ly( x = a$x,
         y = a$y,
         z = matrix(a$z, nrow = length(a$y)),
         type="contour")

plot_ly( z = matrix(PmFor$Temp584m),
         type="contour")

filled.contour(sort(PmFor$Longitude),
               sort(PmFor$Latitude),
               PmFor$Temp584m)

ggplot(PmFor, aes(Longitude, Latitude, z = Temp584m)) +
  geom_contour_filled()

x = rand(15,1)
y = rand(15,1)
z = rand(15,1)

a = interp(x, y, z)

plot_ly(x = a$x,
        y = a$y,
        z = matrix(a$z, nrow = length(a$y), byrow = TRUE),
        type = "contour")
# https://www.r-statistics.com/2016/07/using-2d-contour-plots-within-ggplot2-to-visualize-relationships-between-three-variables/
# 
# data(mtcars)
# data.loess <- loess(qsec ~ wt * hp, data = mtcars)
# # Create a sequence of incrementally increasing (by 0.3 units) values for both wt and hp
# xgrid <-  seq(min(mtcars$wt), max(mtcars$wt), 0.3)
# ygrid <-  seq(min(mtcars$hp), max(mtcars$hp), 0.3)
# # Generate a dataframe with every possible combination of wt and hp
# data.fit <-  expand.grid(wt = xgrid, hp = ygrid)
# # Feed the dataframe into the loess model and receive a matrix output with estimates of
# # acceleration for each combination of wt and hp
# mtrx3d <-  predict(data.loess, newdata = data.fit)
# # Abbreviated display of final matrix
# mtrx3d[1:4, 1:4]

modLoess <- loess(Temp584m ~ Longitude * Latitude, data = PmFor)

xgrid <- seq(min(PmFor$Longitude), max(PmFor$Longitude), 0.01)
ygrid <- seq(min(PmFor$Latitude), max(PmFor$Latitude), 0.01)

data.fit <- expand.grid(Longitude = xgrid, Latitude = ygrid)

mtrx3D <- predict(modLoess, newdata = data.fit)
mtrx3D[1:4, 1:4]

contour(x = xgrid, y = ygrid, z = mtrx3D, xlab = 'Longitude', ylab = 'Latitude')


##ggplot contour map

# Transform data to long form
mtrx.melt <- reshape::melt(mtrx3D, id.vars = c('Longitude', 'Latitude'), measure.vars = 'qsec')
names(mtrx.melt) <- c('Longitude', 'Latitude', 'Temp584m')
# Return data to numeric form
mtrx.melt$Longitude <- as.numeric(str_sub(mtrx.melt$Longitude,
                                          str_locate(mtrx.melt$Longitude, '=')[1,1] + 1))
mtrx.melt$Latitude <- as.numeric(str_sub(mtrx.melt$Latitude,
                                         str_locate(mtrx.melt$Latitude, '=')[1,1] + 1))

head(mtrx.melt)

 ggplot(mtrx.melt, aes(x = Longitude, y = Latitude, z = Temp584m)) +
         stat_contour(geom = 'polygon', aes(fill = ..level..), colour='red') +
         geom_tile(aes(fill = Temp584m)) + # color based on temp values 
         stat_contour(bins = 10, colour='white') +
         xlab('Longitude)') +
         ylab('Latitude') +
         guides(fill = guide_colorbar(title = 'Temp584m (°C)')) +
         theme_bw() 
         scale_fill_discrete(colors = c("#33cc33","#99ff66","#ff6666","#ff0000"))

plot_ly(mtrx.melt, x = ~Longitude, y = ~Latitude, z = ~Temp584m, type = "contour", 
        width = 600, height = 500)
         
         
         
         
         
         

```



```{r echo=FALSE, eval=FALSE}
# ## What is the depth distribution for each type of non-foraging whale group?
#   * Looking more closely at how the different non-foraging groups were distributed across depth, the histograms show that the slow clicks occur most frequently at depths deeper than 4000 m. The groups with codas, and therefore females and juveniles, also occur in the deepest possible depths with only ~1/3 occurring in depths less than 3000 m. It's possible that some of these groups (both slow clicks and codas) may in fact have also been in the process of foraging but not actively foraging at the time of detection.
# 
# Since the non-foraging model only selected depth along with the spatial smoother, this examines the depth distribution for whale groups that included slow clicks (male sperm whales) and codas (males, females, juveniles). These groups were not mutually exclusive. There were more groups with slow clicks than codas overall. 
# 
# The partial effect plot for depth predicted maximum encounter rates at the deepest and shallowest depths with a minimum near 3000-3500 m depth. This corresponds to the depth distributions shown in the histograms for each non-foraging click group. Both groups spanned most depths, but fewer groups occurred near 3000 m in general, corresponding to the model predictions.
```
```{r echo=FALSE, eval=FALSE, fig.height=6, fig.cap='Depth distributions for non-foraging whale groups by click type. The distribution are similar. This is partly due to some overlap between groups, but also shows these groups occur primarily in depths greater than 3500 m.', fig.dim=c(8,6)}
#Check out whos at what depth for the non-foragers
PmNonFor = rbind(trainNonFor, testNonFor)
PmSlow <- dplyr::filter(PmNonFor, nonfor == 1 & grepl('s', click_code))
# nrow(PmSlow)
PmSlow <- PmSlow[, c(1, 54, 2, 4, 34)]

PmCoda <- dplyr::filter(PmNonFor, nonfor == 1 & grepl('cd', click_code))
# nrow(PmCoda)
PmCoda <- PmCoda[, c(1, 54, 2, 4, 34)]

mlay <- rbind(c(1,1), c(2,3))
# print(mlay)
# layout(mlay)
# layout.show(4)

layout(mlay)
par(mar=c(4,4,1,1))
# par(mfrow = c(2,2), mar=c(4,4,1,2))

plot(trainNonFor$Depth, trainNonFor$pa, ylab="Non-Foraging Groups per cell", xlab="Depth")

hist(PmSlow$Depth, breaks = 20, main= 'Slow Clicks', xlab='Depth (m)')
hist(PmCoda$Depth, breaks = 10, main = 'Codas', xlab='Depth (m)')
```


```{r echo=FALSE}
PmFor2 <- subset(PmFor, pa >0)
# par(mfrow= c(2,2))
# hist(PmFor2$temp600_sd.r, main = '')
# plot(PmFor$temp600_sd.r, PmFor$pa, ylab="Foraging Groups per cell", xlab="SD Temp at 584m")
# hist(PmFor2$eke600m.r, main ='')
# plot(PmFor$eke600m.r, PmFor$pa, ylab="Foraging Groups per cell", xlab="EKE at 584m")

modTemp584sd <- gam(temp600_sd.r ~ s(Longitude, Latitude, k=10) , data = PmFor, select = TRUE, method = "REML") 
summary(modTemp584sd)
gam.check(modTemp584sd)

modEKE584 <- gam(eke600m.r ~  s(Longitude, Latitude, k=10) , data = PmFor, select = TRUE, method = "REML") 
summary(modEKE584)
gam.check(modEKE584)

mlay <- rbind(c(2,3),c(1,1))
# print(mlay)
# layout(mlay)
# layout.show(1)

layout(mlay)
par(mar=c(4,4,1,1))

### TempSD 584m Plots####
plot(modTemp584sd, select = 1, scheme = 2, lwd = 2, main = 'Temp at 584mSD ~ LAT:LON, k=10') #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot
points(whales$Longitude, whales$Latitude, pch = 10, col='mediumslateblue', lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)
hist(PmFor2$temp600_sd.r, main = '', xlab="SD Temp at 584m")
plot(PmFor$temp600_sd.r, PmFor$pa, ylab="Foraging Groups per cell", xlab="SD Temp at 584m")


### EKE 584m Plots####
plot(modEKE584, select = 1, scheme = 2, lwd = 2, main = 'EKE at 584m ~ LAT:LON, k=10') #select the first smoother, select = 1
hawaiiMap <- readRDS(here::here(paste0('data/hawaiiMap.rda')))
hawaiiMap$Longitude2 = ifelse(hawaiiMap$Longitude<1, hawaiiMap$Longitude+360, hawaiiMap$Longitude)
whales <- subset(trainHunt, pa >0) #get points for whales to plot
points(whales$Longitude, whales$Latitude, pch = 10, col='mediumslateblue', lwd=1.25, cex=1)
points(hawaiiMap$Longitude2, hawaiiMap$Latitude, pch=20, cex = 0.5)
hist(PmFor2$eke600m.r, main ='')
plot(PmFor$eke600m.r, PmFor$pa, ylab="Foraging Groups per cell", xlab="EKE at 584m")

```


```{r echo=FALSE}
## Add SD of Temp600 to Foraging Mods####
modSuper <- gam(pa ~  s(Longitude, Latitude, k=10) + s(depth.r, k=3) + s(distland.r, k=3) + s(distseamt.r, k=3) +  s(SST, k=3) + s(log.Chla, k=3) + s(Temp584m, k=3) + s(SSH, k=3) + s(SSHsd, k=3) + s(log.eke, k=3) + s(log.wp, k=3) + offset(log.effort)
                + s(temp600_sd.r, k=3) + s(eke600m.r, k=3), data = trainHunt, family = nb,
                link = 'log', select = TRUE, method = "REML")
summary(modSuper) 

modForage2 <- gam(pa ~  s(Longitude, Latitude, k=10) + s(Temp584m, k=3) + s(log.Chla, k=3) + s(SSHsd, k=3) + s(SST, k=3) +offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML") #remove qtr and that makes sst non-sig
summary(modForage2) 

modSuper3 <- gam(pa ~  s(Longitude, Latitude, k=10) +
                    s(temp600_sd.r, k=3) + offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modSuper3) 

modSuper4 <- gam(pa ~  s(Longitude, Latitude, k=10) +
                    s(eke600m.r, k=3) + offset(log.effort) , data = trainHunt, family = nb, link = 'log', select = TRUE, method = "REML")
summary(modSuper4) 
```
